% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{appendVariables}
\alias{appendVariables}
\alias{appendVariables,VirtualReferenceData,data.frame,character-method}
\title{Append new columns to stored data}
\usage{
appendVariables(object, data, var.names)

\S4method{appendVariables}{VirtualReferenceData,data.frame,character}(object,
  data, var.names)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data}

\item{var.names}{column names of data in both incomming data}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.

\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Adds new columns to stored data
The number of rows of incoming data has to be the same as number of
rows currently stored.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualReferenceData,data = data.frame,var.names = character}: Append new columns to stored data

Adds new columns to stored data
The number of rows of incoming data has to be the same as number of
rows currently stored.
}}

