% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{attachTransformation}
\alias{attachTransformation}
\title{Attach transformation to the object}
\usage{
attachTransformation(object, transformation, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformation}{objects of clas "VirtualReferenceDataTransformation"}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}

