% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\docType{methods}
\name{attachTransformations}
\alias{attachTransformations}
\alias{attachTransformations,VirtualTransformationsHandler,list-method}
\title{Attach transformations to the object}
\usage{
attachTransformations(object, transformations, replace = TRUE)

\S4method{attachTransformations}{VirtualTransformationsHandler,list}(object,
  transformations, replace = TRUE)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformations}{list of objects of class "VirtualReferenceDataTransformation"}

\item{replace}{"logical" replace transformation(s) if already attached.}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.

\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformationsHandler,transformations = list}: Attach transformations to the object

Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}}

