% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_traders.r
\docType{class}
\name{AvailableTraders-class}
\alias{AvailableTraders-class}
\title{Contrete S4 class providing a dummy mechanism for
analysis blocks to query the available
trader ids.}
\description{
dataRequest implementation supplies the
values from a static .obtainValues
implementation.
}
\details{
Inherits from "VirtualAvailableValues
}

