% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_portfolio.r
\docType{methods}
\name{buildPortfolioHistory,VirtualIndexPortfolio-method}
\alias{buildPortfolioHistory,VirtualIndexPortfolio-method}
\title{Pool data required for portfolio}
\usage{
\S4method{buildPortfolioHistory}{VirtualIndexPortfolio}(object, start, end)
}
\arguments{
\item{object}{object of class 'VirtualIndexPortfolio'.}

\item{start}{Date start of data range.}

\item{end}{Date end of data range.}
}
\value{
\code{object} object of class 'VirtualIndexPortfolio'.
}
\description{
Pools data from data source specific to derived class implementation
Also computes portfolio weights if necessary
}

