% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\name{buildPortfolioHistory}
\alias{buildPortfolioHistory}
\title{Pool data required for portfolio}
\usage{
buildPortfolioHistory(object, start, end)
}
\arguments{
\item{object}{object of class 'Portfolio'.}

\item{start}{Date start of data range.}

\item{end}{Date end of data range.}
}
\value{
\code{object} object of class 'Portfolio'.
}
\description{
Pools data from data source specific to derived class implementation
Also computes portfolio weights if necessary
}

