% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{CompoundImpliedFactorReturnsTransformationComputation-class}
\alias{CompoundImpliedFactorReturnsTransformationComputation-class}
\title{computes compounded factor return}
\description{
Adds cumulant columns.
}
\details{
Inherits from "VirtualTransformationComputation"
}

