% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_return_state.r
\docType{methods}
\name{computeImpliedFactorReturnsState}
\alias{computeImpliedFactorReturnsState}
\alias{computeImpliedFactorReturnsState,ImpliedFactorReturnsState-method}
\title{Apply transformations in sequence in order to compute the
factor return state.}
\usage{
computeImpliedFactorReturnsState(object)

\S4method{computeImpliedFactorReturnsState}{ImpliedFactorReturnsState}(object)
}
\arguments{
\item{object}{object of class 'ImpliedFactorReturnsState'.}
}
\value{
\code{object} object of class 'ImpliedFactorReturnsState'.

\code{object} object of class 'ImpliedFactorReturnsState'.
}
\description{
Apply transformations in sequence in order to compute the
factor return state.
}
\section{Methods (by class)}{
\itemize{
\item \code{ImpliedFactorReturnsState}: Compute factor return state

Applies transformations in the transformation handler
}}

