% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{computeTransformation}
\alias{computeTransformation}
\title{Trigger computation}
\usage{
computeTransformation(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{object} object object of class 'VirtualTransformationComputation'.
}
\description{
triggers computation and copies computed data to output datastore.
input has to be set using setInputData(object, data) prior to triggering
this function otherwise will return error
}

