% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_portfolio.r
\docType{methods}
\name{dataRequest,StrategyPortfolio,data.frame-method}
\alias{dataRequest,StrategyPortfolio,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{StrategyPortfolio,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'StrategyPortfolio'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'StrategyPortfolio'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

