% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client_test_class.r
\docType{methods}
\name{dataRequest,TestDataStoreClient,data.frame-method}
\alias{dataRequest,TestDataStoreClient,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{TestDataStoreClient,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'TestDataStoreClient'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'TestDataStoreClient'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

