% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client.r
\docType{methods}
\name{dataRequest,VirtualDataStoreClient,data.frame-method}
\alias{dataRequest,VirtualDataStoreClient,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{VirtualDataStoreClient,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class 'VirtualDataStoreClient'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'VirtualDataStoreClient'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

