% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{methods}
\name{dataRequest,VirtualRiskModelObjectstoreClient,data.frame-method}
\alias{dataRequest,VirtualRiskModelObjectstoreClient,data.frame-method}
\title{Request data from data source}
\usage{
\S4method{dataRequest}{VirtualRiskModelObjectstoreClient,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelObjectstoreClient'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'VirtualRiskModelObjectstoreClient'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

