% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{dataRequest}
\alias{dataRequest}
\title{Request data from data source}
\usage{
dataRequest(object, key_values, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{key_values}{data.frame with keys specifying data query.}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Generic method to request data from data source.
Needs to be implemented in derived classes to work
}

