% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_portfolio_transformations.r
\docType{class}
\name{DaysSinceLastFlatTransformation-class}
\alias{DaysSinceLastFlatTransformation-class}
\title{computes number of days since last flat}
\description{
Adds column with number of days since last flat
for each position
Inherits from "VirtualReferenceDataTransformation"
uses "DaysSinceLastFlatTransformationComputation"
}

