% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{class}
\name{DaysSinceLastFlatTransformationComputation-class}
\alias{DaysSinceLastFlatTransformationComputation-class}
\title{computes number of days since last flat}
\description{
Adds column with number of days since last flat
for each position.
Inherits from "VirtualTransformationComputation"
}

