% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dealer_data.r
\docType{class}
\name{DealerData-class}
\alias{DealerData-class}
\title{Concrete S4 class for handling Dealer Data.}
\description{
Test class for testing VirtualDataStoreClient
functionality. Do not use.
}
\details{
Inherits from "VirtualDataStoreClient"
}

