% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{executeSQLQuery}
\alias{executeSQLQuery}
\title{Executes SQL query strings}
\usage{
executeSQLQuery(object, key_values, ...)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}

\item{key_values}{data_frame with query keys. If missing will use
previously set keys or returns an error.}
}
\value{
\code{ret_val} data.frame with result of the query
}
\description{
Executes previously prepared SQL queries and returns result
}

