% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traderid_key_expander.r
\docType{methods}
\name{expandGGPlot}
\alias{expandGGPlot}
\alias{expandGGPlot,VirtualTradeIDKeyExpander,gg-method}
\title{Modify the ggplot if desired}
\usage{
expandGGPlot(object, ggplot)

\S4method{expandGGPlot}{VirtualTradeIDKeyExpander,gg}(object, ggplot)
}
\arguments{
\item{object}{object of class "VirtualTradeIDKeyExpander"}

\item{ggplot}{object of class "gg"}
}
\value{
\code{ggplot} object of class "gg"

\code{ggplot} object of class "gg"
}
\description{
Modify the ggplot if desired
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTradeIDKeyExpander,ggplot = gg}: Modify the ggplot if desired
}}

