% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data.r
\docType{class}
\name{FactorCorrelationData-class}
\alias{FactorCorrelationData-class}
\title{Concrtete S4 class handling Factor Correlation data.}
\description{
Implements storage for Factor Correlation and access
to Factor Correlation data via Risk Model Objectstore
}
\details{
Inherits from "VirtualFactorCorrelationData" and "VirtualRiskModelObjectstoreClient"
}

