% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data.r
\docType{class}
\name{FactorVarianceData-class}
\alias{FactorVarianceData-class}
\title{Concrtete S4 class handling Factor Variance data.}
\description{
Implements storage for Factor Variance and access
to Factor Variance data via Risk Model Objectstore
}
\details{
Inherits from "VirtualFactorVarianceData" and "VirtualRiskModelObjectstoreClient"
}

