% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{getAttachedTransformations}
\alias{getAttachedTransformations}
\title{Get list of attached transformations to the Object}
\usage{
getAttachedTransformations(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}
}
\value{
\code{transformations} list of "VirtualReferenceDataTransformation" objects attached.
}
\description{
Returns list of names of classes of transformations that
have been attached to object
}

