% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values_data_handler.r
\docType{methods}
\name{getAvailableValuesObject}
\alias{getAvailableValuesObject}
\alias{getAvailableValuesObject,VirtualAvailableValuesDataHandler-method}
\title{Get available_Values stored in object}
\usage{
getAvailableValuesObject(object)

\S4method{getAvailableValuesObject}{VirtualAvailableValuesDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualAvailableValuesDataHandler"}
}
\value{
\code{available_Values} object of class "VirtualAVailableValues"

\code{available_Values} object of class "VirtualAvailableValues"
}
\description{
Returns available_Values object of class "VirtualAvailableValues"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAvailableValuesDataHandler}: Get available_Values stored in object

Returns available_Values object of class "VirtualAvailableValues"
}}

