% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getComputation}
\alias{getComputation}
\alias{getComputation,VirtualTransformation-method}
\title{Get computation object}
\usage{
getComputation(object)

\S4method{getComputation}{VirtualTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{compute} function used for computation of specific transformation

\code{compute} function used for computation of specific transformation
}
\description{
Returns computation object encapsulated in Transformation .
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformation}: Get computation object

Returns computation object encapsulated in Transformation .
}}

