% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getComputationFunction}
\alias{getComputationFunction}
\title{Get code of computation function}
\usage{
getComputationFunction(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{compute} function used for computation of specific transformation
}
\description{
Returns function object used for computation
}

