% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getComputationInput}
\alias{getComputationInput}
\alias{getComputationInput,VirtualTransformation-method}
\title{Get value of computation input}
\usage{
getComputationInput(object)

\S4method{getComputationInput}{VirtualTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}
}
\value{
\code{data} data.frame with transformation computation input

\code{data} data.frame with transformation computation input
}
\description{
Returns internal input datastore of computation object .
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformation}: Get value of computation input

Returns internal input datastore of computation object .
}}

