% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getComputationOutput}
\alias{getComputationOutput}
\alias{getComputationOutput,VirtualTransformation-method}
\title{Get value of computation output}
\usage{
getComputationOutput(object)

\S4method{getComputationOutput}{VirtualTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}
}
\value{
\code{data} data.frame with transformation computation output

\code{data} data.frame with transformation computation output
}
\description{
Returns internal output datastore of computation object .
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformation}: Get value of computation output

Returns internal output datastore of computation object .
}}

