% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getComputedVariablesNames}
\alias{getComputedVariablesNames}
\alias{getComputedVariablesNames,VirtualBaseTransformation-method}
\title{Get names of computed columns}
\usage{
getComputedVariablesNames(object)

\S4method{getComputedVariablesNames}{VirtualBaseTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{computed_colnms} character vector of names of computed columns

\code{computed_colnms} character vector of names of computed columns
}
\description{
Returns Names of the variables that are generated by computation object.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualBaseTransformation}: Get names of computed columns

Returns Names of the variables that are generated by computation object.
}}

