% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{getDataSourceClientColumnNameMap}
\alias{getDataSourceClientColumnNameMap}
\title{Retreive datasource column name map}
\usage{
getDataSourceClientColumnNameMap(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{column_name_map} hash column name map of the datasource to output names
}
\description{
Returns hash with mapping of datasource column names
to reference data class column names
}

