% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{getDataSourceQueryKeyColumnNames}
\alias{getDataSourceQueryKeyColumnNames}
\alias{getDataSourceQueryKeyColumnNames,VirtualDataSourceClient-method}
\title{Get query key column names}
\usage{
getDataSourceQueryKeyColumnNames(object)

\S4method{getDataSourceQueryKeyColumnNames}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{key_cols} character vector with list of key column names

\code{key_cols} character vector with list of key column names
}
\description{
Private method to return key column names
that are used to query raw data source
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Get query key column names

Private method to return key column names
that are used to query raw data source
}}

