% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{getDataSourceQueryKeyValues}
\alias{getDataSourceQueryKeyValues}
\alias{getDataSourceQueryKeyValues,VirtualDataSourceClient-method}
\title{Retreive values of query keys}
\usage{
getDataSourceQueryKeyValues(object)

\S4method{getDataSourceQueryKeyValues}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_non_na_cols} logical

\code{has_non_na_cols} logical
}
\description{
Get values of key values datatable
with keys used to  match in datasource query
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Retreive values of query keys

Get values of key values datatable
with keys used to  match in datasource query
}}

