% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{getDataSourceReturnColumnNames}
\alias{getDataSourceReturnColumnNames}
\title{Get returned source data column names}
\usage{
getDataSourceReturnColumnNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{values} character vector with list of returned source column names
}
\description{
Returns names of the columns that will be returned from data source
}

