% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client.r
\name{getDataStoreName}
\alias{getDataStoreName}
\title{Get source datastore name}
\usage{
getDataStoreName(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataStoreClient'.}
}
\value{
\code{datastore_name} character
}
\description{
Returns name of the datastore which will be querried when asked to fill data.
}

