% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\name{getEndDate}
\alias{getEndDate}
\title{Get end date of portfolio timespan}
\usage{
getEndDate(object, ...)
}
\arguments{
\item{object}{object of class 'Portfolio'.}
}
\value{
\code{end} Date end of data range.
}
\description{
Get end date of portfolio timespan
}

