% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_data_handler.r
\docType{methods}
\name{getEventDataObject}
\alias{getEventDataObject}
\alias{getEventDataObject,VirtualEventDataHandler-method}
\title{Get event_data stored in object}
\usage{
getEventDataObject(object)

\S4method{getEventDataObject}{VirtualEventDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualEventDataHandler"}
}
\value{
\code{event_data} object of class "VirtualEventData"

\code{event_data} object of class "VirtualEventData"
}
\description{
Returns event_data object of class "VirtualEventData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualEventDataHandler}: Get event_data stored in object

Returns event_data object of class "VirtualEventData"
}}

