% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_expander.r
\docType{methods}
\name{getExpanderKeys}
\alias{getExpanderKeys}
\alias{getExpanderKeys,VirtualKeyExpander-method}
\title{Get key_values}
\usage{
getExpanderKeys(object)

\S4method{getExpanderKeys}{VirtualKeyExpander}(object)
}
\arguments{
\item{object}{object of class "VirtualKeyExpander"}
}
\value{
\code{key_values} object of class "data.frame"

\code{key_values} object of class "data.frame"
}
\description{
Get key_values
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualKeyExpander}: Get key_values
}}

