% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trade_data_handler.r
\docType{methods}
\name{getExtendedTradeDataObject}
\alias{getExtendedTradeDataObject}
\alias{getExtendedTradeDataObject,VirtualExtendedTradeDataHandler-method}
\title{Get ex_trade data stored in object}
\usage{
getExtendedTradeDataObject(object)

\S4method{getExtendedTradeDataObject}{VirtualExtendedTradeDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualExtendedTradeDataHandler"}
}
\value{
\code{ex_trade_data} object of class "VirtualExtendedTradeData"

\code{ex_trade_data} object of class "VirtualExtendedTradeData"
}
\description{
Returns ex_trade_data object of class "VirtualExtendedTradeData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualExtendedTradeDataHandler}: Get ex_trade data stored in object

Returns ex_trade_data object of class "VirtualExtendedTradeData"
}}

