% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data_handler.r
\name{getFactorCorrelationDataObject}
\alias{getFactorCorrelationDataObject}
\title{Get factor_correlation data stored in object}
\usage{
getFactorCorrelationDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualFactorCorrelationDataHandler"}
}
\value{
\code{factor_correlation} object of class "VirtualFactorCorrelationData"
}
\description{
Returns factor_correlation object of class "VirtualFactorCorrelationData"
}

