% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_exposure_data_handler.r
\name{getFactorExposureDataObject}
\alias{getFactorExposureDataObject}
\title{Get factor_exposure_data stored in object}
\usage{
getFactorExposureDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualFactorExposureDataHandler"}
}
\value{
\code{factor_exposure_data} object of class "VirtualFactorExposureData"
}
\description{
Returns factor_exposure_data object of class "VirtualFactorExposureData"
}

