% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data_handler.r
\name{getFactorVarianceDataObject}
\alias{getFactorVarianceDataObject}
\title{Get factor_variance stored in object}
\usage{
getFactorVarianceDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualFactorVarianceDataHandler"}
}
\value{
\code{factor_variance} object of class "VirtualFactorVarianceData"
}
\description{
Returns factor_variance object of class "VirtualFactorVarianceData"
}

