% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{getForceUniqueRows}
\alias{getForceUniqueRows}
\alias{getForceUniqueRows,VirtualReferenceData-method}
\title{Are stored rows forced to be unique}
\usage{
getForceUniqueRows(object)

\S4method{getForceUniqueRows}{VirtualReferenceData}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{unique_rows} logical, are stored rows unique

\code{unique_rows} logical, are stored rows unique
}
\description{
Returns logical value indicating if stored reference data
rows are forced to be unique.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceData}: Are stored rows forced to be unique

Returns logical value indicating if stored reference data
rows are forced to be unique.
}}

