% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data_handler.r
\name{getImpliedFactorReturnsDataObject}
\alias{getImpliedFactorReturnsDataObject}
\title{Get implied_factor_returns stored in object}
\usage{
getImpliedFactorReturnsDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualImpliedFactorReturnsDataHandler"}
}
\value{
\code{implied_factor_returns} object of class "VirtualImpliedFactorReturnsData"
}
\description{
Returns implied_factor_returns object of class "VirtualImpliedFactorReturnsData"
}

