% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_portfolio.r
\docType{methods}
\name{getIndexTicker}
\alias{getIndexTicker}
\alias{getIndexTicker,VirtualIndexPortfolio-method}
\title{Returns IndexTicker name}
\usage{
getIndexTicker(object)

\S4method{getIndexTicker}{VirtualIndexPortfolio}(object)
}
\arguments{
\item{object}{object of class 'VirtualIndexPortfolio'.}
}
\value{
\code{index_ticker} integer.
}
\description{
Returns IndexTicker name
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualIndexPortfolio}: Returns TraderID
}}

