% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getInputData}
\alias{getInputData}
\title{Get value of computation input}
\usage{
getInputData(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{data} data.frame data stored in input
}
\description{
Returns data from internal input datastore.
}

