% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data_handler.r
\name{getInstrumentBetasDataObject}
\alias{getInstrumentBetasDataObject}
\title{Get instrument_betas stored in object}
\usage{
getInstrumentBetasDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualInstrumentBetasDataHandler"}
}
\value{
\code{instrument_betas} object of class "VirtualInstrumentBetasData"
}
\description{
Returns instrument_betas object of class "VirtualInstrumentBetasData"
}

