% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_residual_returns_data_handler.r
\docType{methods}
\name{getInstrumentResidualReturnsDataObject}
\alias{getInstrumentResidualReturnsDataObject}
\alias{getInstrumentResidualReturnsDataObject,VirtualInstrumentResidualReturnsDataHandler-method}
\title{Get instrument_residual_returns stored in object}
\usage{
getInstrumentResidualReturnsDataObject(object)


  \S4method{getInstrumentResidualReturnsDataObject}{VirtualInstrumentResidualReturnsDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualInstrumentResidualReturnsDataHandler"}
}
\value{
\code{instrument_residual_returns} object of class "VirtualInstrumentResidualReturnsData"

\code{instrument_residual_returns} object of class "VirtualInstrumentResidualReturnsData"
}
\description{
Returns instrument_residual_returns object of class "VirtualInstrumentResidualReturnsData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualInstrumentResidualReturnsDataHandler}: Get instrument_residual_returns stored in object

Returns instrument_residual_returns object of class "VirtualInstrumentResidualReturnsData"
}}

