% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data_handler.r
\docType{methods}
\name{getMarketDataObject}
\alias{getMarketDataObject}
\alias{getMarketDataObject,VirtualMarketDataHandler-method}
\title{Get market_data stored in object}
\usage{
getMarketDataObject(object)

\S4method{getMarketDataObject}{VirtualMarketDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualMarketDataHandler"}
}
\value{
\code{market_data} object of class "VirtualMarketData"

\code{market_data} object of class "VirtualMarketData"
}
\description{
Returns market_data object of class "VirtualMarketData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualMarketDataHandler}: Get market_data stored in object

Returns market_data object of class "VirtualMarketData"
}}

