% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data_handler.r
\name{getMarketStyleDataObject}
\alias{getMarketStyleDataObject}
\title{Get market_style stored in object}
\usage{
getMarketStyleDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualMarketStyleDataHandler"}
}
\value{
\code{market_style} object of class "VirtualMarketStyleData"
}
\description{
Returns market_style object of class "VirtualMarketStyleData"
}

