% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{getNonNAColumnNames}
\alias{getNonNAColumnNames}
\title{Get non - NA's column names}
\usage{
getNonNAColumnNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{non_na_cols} character vector with list of returned column names
}
\description{
Returns column names that are required to have non-na values
rows where these columns have NA's will be removed
}

