% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getOutputData}
\alias{getOutputData}
\alias{getOutputData,VirtualTransformationComputation-method}
\title{Get value of computation output}
\usage{
getOutputData(object)

\S4method{getOutputData}{VirtualTransformationComputation}(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{data} data.frame data stored in output

\code{data} data.frame data stored in output
}
\description{
Returns data from internal output datastore.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformationComputation}: Get value of computation output

Returns data from internal output datastore.
}}

