% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getOutputVariablesNames}
\alias{getOutputVariablesNames}
\title{Get names of output columns}
\usage{
getOutputVariablesNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{output_colnms} character vector of names of output columns
}
\description{
Returns Names of the variables that are returned by computation object.
}

