% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_data_handler.r
\name{getPortfolioDataObject}
\alias{getPortfolioDataObject}
\title{Get portfolio stored in object}
\usage{
getPortfolioDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualPortfolioDataHandler"}
}
\value{
\code{portfolio} object of class "Portfolio"
}
\description{
Returns portfolio object of class "Portfolio"
}

