% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data_handler.r
\docType{methods}
\name{getPositionDataObject}
\alias{getPositionDataObject}
\alias{getPositionDataObject,VirtualPositionDataHandler-method}
\title{Get position_data stored in object}
\usage{
getPositionDataObject(object)

\S4method{getPositionDataObject}{VirtualPositionDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualPositionDataHandler"}
}
\value{
\code{position_data} object of class "VirtualPositionData"

\code{position_data} object of class "VirtualPositionData"
}
\description{
Returns position_data object of class "VirtualPositionData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualPositionDataHandler}: Get position_data stored in object

Returns position_data object of class "VirtualPositionData"
}}

