% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmodel_client.r
\name{getPPModelClass}
\alias{getPPModelClass}
\title{Get Post Processing Model Class}
\usage{
getPPModelClass(object, ...)
}
\arguments{
\item{object}{object of class "VirtualPPModelClient"}
}
\description{
Returns post processing model used to assembly data.
}

