% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data_handler.r
\name{getPriceDataObject}
\alias{getPriceDataObject}
\title{Get price_data stored in object}
\usage{
getPriceDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualPriceDataHandler"}
}
\value{
\code{price_data} object of class "VirtualPriceData"
}
\description{
Returns price_data object of class "VirtualPriceData"
}

