% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r, R/referencedata_transformation.r
\name{getReferenceData}
\alias{getReferenceData}
\title{get stored data}
\usage{
getReferenceData(object, ...)

getReferenceData(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{object}{object of class 'VirtualReferenceDataTransformation'.}
}
\value{
\code{data} data.frame with stored data..

\code{ref_data} object of class 'VirtualReferenceData'.
}
\description{
Returns data.frame with stored data.

Returns VirtualReferenceData object on which transformation is beeing done
}

