% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\name{getReferenceDataObject}
\alias{getReferenceDataObject}
\title{Get stored VirtualReferenceData Object}
\usage{
getReferenceDataObject(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceDataTransformation'.}
}
\value{
\code{ref_data} object of class 'VirtualReferenceData'.
}
\description{
Returns VirtualReferenceData object on which transformation is beeing done
}

