% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceobject.r
\docType{methods}
\name{getRequiredVariablesNames}
\alias{getRequiredVariablesNames}
\alias{getRequiredVariablesNames,VirtualReferenceObject-method}
\title{Get names of required variables stored in data}
\usage{
getRequiredVariablesNames(object)

\S4method{getRequiredVariablesNames}{VirtualReferenceObject}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceObject'.}
}
\value{
\code{required_colnms} character vector, list of required column names

\code{required_colnms} character vector, list of required column names
}
\description{
Get names of required variables stored in data
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceObject}: Get names of required variables stored in data
}}

