% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{methods}
\name{getRiskModelComponentName}
\alias{getRiskModelComponentName}
\alias{getRiskModelComponentName,VirtualRiskModelDataSourceClient-method}
\title{Get Risk Model Component Name}
\usage{
getRiskModelComponentName(object)

\S4method{getRiskModelComponentName}{VirtualRiskModelDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelDataSourceClient'.}
}
\value{
\code{component} 'character', name of the component beeing accessed.

\code{component} 'character', name of the component beeing accessed.
}
\description{
Returns name of the component that given class is accessing
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModelDataSourceClient}: Get Risk Model Component Name

Returns name of the component that given class is accessing
}}

