% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelFactorNames}
\alias{getRiskModelFactorNames}
\title{Get Risk Model Factor Names}
\usage{
getRiskModelFactorNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_factors} 'character', vector of factor names
}
\description{
Returns names of factors included in the model
}

