% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelLookback,VirtualRiskModelHandler-method}
\alias{getRiskModelLookback,VirtualRiskModelHandler-method}
\title{Get Risk Model Lookback}
\usage{
\S4method{getRiskModelLookback}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{lookback} 'integer'
}
\description{
Returns model lookback value in days. Lookback is number of
calendar days over which model is calculated.
}

