% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelLookback}
\alias{getRiskModelLookback}
\title{Get Risk Model Lookback}
\usage{
getRiskModelLookback(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{lookback} 'integer'
}
\description{
Returns model lookback value in days. Lookback is number of
calendar days over which model is calculated.
}

