% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelName,VirtualRiskModelHandler-method}
\alias{getRiskModelName,VirtualRiskModelHandler-method}
\title{Get Risk Model Name}
\usage{
\S4method{getRiskModelName}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{model_name} 'character', name of the model.
}
\description{
Returns name of Risk Model
}

