% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelName}
\alias{getRiskModelName}
\title{Get Risk Model Name}
\usage{
getRiskModelName(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_name} 'character', name of the model.
}
\description{
Returns name of Risk Model
}

