% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\name{getRiskModelObject}
\alias{getRiskModelObject}
\title{Get Risk Model Object}
\usage{
getRiskModelObject(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{risk_model} object of class "VirtualRiskModel".
}
\description{
Returns risk model object
}

