% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{methods}
\name{getRiskModelObjectstoreComponentName}
\alias{getRiskModelObjectstoreComponentName}
\alias{getRiskModelObjectstoreComponentName,VirtualRiskModelObjectstoreClient-method}
\title{Get Risk Model Component Name}
\usage{
getRiskModelObjectstoreComponentName(object)


  \S4method{getRiskModelObjectstoreComponentName}{VirtualRiskModelObjectstoreClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelObjectstoreClient'.}
}
\value{
\code{component} 'character', name of the component beeing accessed.

\code{component} 'character', name of the component beeing accessed.
}
\description{
Returns name of the component that given class is accessing
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModelObjectstoreClient}: Get Risk Model Component Name

Returns name of the component that given class is accessing
}}

