% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelPrefix,VirtualRiskModelHandler-method}
\alias{getRiskModelPrefix,VirtualRiskModelHandler-method}
\title{Get Risk Model Prefix}
\usage{
\S4method{getRiskModelPrefix}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{model_prefix} 'character', model prefix.
}
\description{
Returns prefix of Risk Model used to construct model store querries
}

