% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelPrefix}
\alias{getRiskModelPrefix}
\title{Get Risk Model Prefix}
\usage{
getRiskModelPrefix(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_prefix} 'character', model prefix.
}
\description{
Returns prefix of Risk Model used to construct model store querries
}

