% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelUniverse,VirtualRiskModelHandler-method}
\alias{getRiskModelUniverse,VirtualRiskModelHandler-method}
\title{Get Risk Model Universe}
\usage{
\S4method{getRiskModelUniverse}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{model_universe} 'character', name of the universe covered by the model
}
\description{
Returns name of universe for which model is computed.
}

