% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{methods}
\name{getRiskModelUniverse}
\alias{getRiskModelUniverse}
\alias{getRiskModelUniverse,VirtualRiskModel-method}
\title{Get Risk Model Universe}
\usage{
getRiskModelUniverse(object)

\S4method{getRiskModelUniverse}{VirtualRiskModel}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_universe} 'character', name of the universe covered by the model

\code{model_universe} 'character', name of the universe covered by the model
}
\description{
Returns name of universe for which model is computed.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModel}: Get Risk Model Universe

Returns name of universe for which model is computed.
}}

