% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{getSQLQueryKeyColumnNames}
\alias{getSQLQueryKeyColumnNames}
\title{Get sql query column names}
\usage{
getSQLQueryKeyColumnNames(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{key_cols} character vector of constructed SQL query strings.
}
\description{
Returns key column names that are used to construct sql querries.
}

