% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{getSQLQueryKeyValues}
\alias{getSQLQueryKeyValues}
\title{Get sql query column names}
\usage{
getSQLQueryKeyValues(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{key_values} data.frame with columns matching key column names()
}
\description{
Returns key values datatable
with keys that will be matched in datastore query
}

