% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\docType{methods}
\name{getStartDate}
\alias{getStartDate}
\alias{getStartDate,Portfolio-method}
\title{Get start date of portfolio timespan}
\usage{
getStartDate(object)

\S4method{getStartDate}{Portfolio}(object)
}
\arguments{
\item{object}{object of class 'Portfolio'.}
}
\value{
\code{start} Date start of data range.

\code{start} Date start of data range.
}
\description{
Get start date of portfolio timespan
}
\section{Methods (by class)}{
\itemize{
\item \code{Portfolio}: Get start date of portfolio timespan
}}

