% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{getStoredNRows}
\alias{getStoredNRows}
\alias{getStoredNRows,VirtualReferenceData-method}
\title{Number of currently stored rows}
\usage{
getStoredNRows(object)

\S4method{getStoredNRows}{VirtualReferenceData}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{stored_rows} integer, list of stored column names

\code{stored_rows} integer, list of stored column names
}
\description{
Returns count of currently stored rows
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceData}: Number of currently stored rows

Returns count of currently stored rows
}}

