% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{getStoredVariablesNames}
\alias{getStoredVariablesNames}
\alias{getStoredVariablesNames,VirtualReferenceData-method}
\title{Get names of stored reference data variables}
\usage{
getStoredVariablesNames(object)

\S4method{getStoredVariablesNames}{VirtualReferenceData}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{column_names} character vector, list of stored column names

\code{column_names} character vector, list of stored column names
}
\description{
Returns vector of column names of stored data
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceData}: Get names of stored reference data variables

Returns vector of column names of stored data
}}

