% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_data_handler.r
\docType{methods}
\name{getStrategyDataObject}
\alias{getStrategyDataObject}
\alias{getStrategyDataObject,VirtualStrategyDataHandler-method}
\title{Get strategy_data stored in object}
\usage{
getStrategyDataObject(object)

\S4method{getStrategyDataObject}{VirtualStrategyDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualStrategyDataHandler"}
}
\value{
\code{strategy_data} object of class "VirtualStrategyData"

\code{strategy_data} object of class "VirtualStrategyData"
}
\description{
Returns strategy_data object of class "VirtualStrategyData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualStrategyDataHandler}: Get strategy_data stored in object

Returns strategy_data object of class "VirtualStrategyData"
}}

