% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data_handler.r
\name{getTradeDataObject}
\alias{getTradeDataObject}
\title{Get trade_data stored in object}
\usage{
getTradeDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualTradeDataHandler"}
}
\value{
\code{trade_data} object of class "VirtualTradeData"
}
\description{
Returns trade_data object of class "VirtualTradeData"
}

