% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_portfolio.r
\name{getTraderID}
\alias{getTraderID}
\title{Returns TraderID}
\usage{
getTraderID(object, ...)
}
\arguments{
\item{object}{object of class 'StrategyPortfolio'.}
}
\value{
\code{trader_id} integer.
}
\description{
Returns TraderID
}

