% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values_data_handler.r
\name{getValueName}
\alias{getValueName}
\title{Getter for the value column name}
\usage{
getValueName(object)
}
\arguments{
\item{object}{object of class 'VirtualAvailableValuesDataHandler'.}
}
\value{
\code{name} "character"
}
\description{
Getter for the value column name
}

