% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{hasFactorColumnNames}
\alias{hasFactorColumnNames}
\alias{hasFactorColumnNames,VirtualDataSourceClient-method}
\title{Are there any columns to  factorize ?}
\usage{
hasFactorColumnNames(object)

\S4method{hasFactorColumnNames}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_factorized_cols} logical

\code{has_factorized_cols} logical
}
\description{
Returns TRUE if there are some columns to factorize
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Are there any columns to  factorize ?

Returns TRUE if there are some columns to factorize
}}

