% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{hasNonNAColumnNames}
\alias{hasNonNAColumnNames}
\title{Are any non - NA's column names defined in object}
\usage{
hasNonNAColumnNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_non_na_cols} logical
}
\description{
Returns true if object has columns that are required to have non-na-values
}

