% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_expander.r
\docType{methods}
\name{hasQueryBeenExpanded}
\alias{hasQueryBeenExpanded}
\alias{hasQueryBeenExpanded,VirtualKeyExpander-method}
\title{Get has_been_expanded}
\usage{
hasQueryBeenExpanded(object)

\S4method{hasQueryBeenExpanded}{VirtualKeyExpander}(object)
}
\arguments{
\item{object}{object of class "VirtualKeyExpander"}
}
\value{
\code{has_been_expanded} object of class "logical"

\code{has_been_expanded} object of class "logical"
}
\description{
Get has_been_expanded
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualKeyExpander}: Get has_been_expanded
}}

