% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data.r
\docType{class}
\name{ImpliedFactorReturnsData-class}
\alias{ImpliedFactorReturnsData-class}
\title{Concrtete S4 class handling Implied Factor Returns data.}
\description{
Implements storage for Implied Factor Returns and access
to Implied Factor Returns data via Risk Model Objectstore
}
\details{
Inherits from "VirtualImpliedFactorReturnsData" and "VirtualRiskModelObjectstoreClient"
}

