% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsMAVGSpreadTransformation-class}
\alias{ImpliedFactorReturnsMAVGSpreadTransformation-class}
\title{computes return mavg spread}
\description{
Adds columns for the spread and spread diff
for each factor
}
\details{
Inherits from "VirtualTransformation"
uses "CompoundImpliedFactorReturnsMAVGSpreadTransformationComputation"
}

