% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsMAVGSpreadTransformationComputation-class}
\alias{ImpliedFactorReturnsMAVGSpreadTransformationComputation-class}
\title{computes spread between moving averages of the compounded factor return}
\description{
Adds columns holding the spread and the diff of the spread
}
\details{
Inherits from "VirtualTransformationComputation"
}

