% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsMAVGTransformation-class}
\alias{ImpliedFactorReturnsMAVGTransformation-class}
\title{computes compound implied factor return}
\description{
Adds columns for the compound return for each
factor.
}
\details{
Inherits from "VirtualTransformation"
uses "CompoundImpliedFactorReturnsTransformationComputation"
}

