% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsMAVGTransformationComputation-class}
\alias{ImpliedFactorReturnsMAVGTransformationComputation-class}
\title{computes moving averages of the compounded factor return}
\description{
Adds columns holding the MAVG values.
}
\details{
Inherits from "VirtualTransformationComputation"
}

