% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsQuartileTransformationComputation-class}
\alias{ImpliedFactorReturnsQuartileTransformationComputation-class}
\title{computes quartile of the compound factor return level}
\description{
Adds columns holding the quartile for each day
}
\details{
Inherits from "VirtualTransformationComputation"
}

