% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_return_state.r
\docType{class}
\name{ImpliedFactorReturnsState-class}
\alias{ImpliedFactorReturnsState-class}
\title{S4 class implementing encoding of
factor return state in terms of the
quartile of the compound return level
and whether the factor is sideways trending
or reverting.}
\description{
This is concrete class handling  computation
of the factor return state.
Inherits from:
 "ImpliedFactorReturnsData"
 "VirtualTransformationsHandler"
}

