% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsStateEncoderTransformation-class}
\alias{ImpliedFactorReturnsStateEncoderTransformation-class}
\title{computes factor state based on substates space}
\description{
Adds columns holding factor state
}
\details{
Inherits from "VirtualReferenceDataTransformation"
uses "ImpliedFactorReturnsStateEncoderTransformationComputation"
}

