% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsTrendIndicatorTransformation-class}
\alias{ImpliedFactorReturnsTrendIndicatorTransformation-class}
\title{computes return quantile}
\description{
Adds columns for the quantile of compounded return
for each factor
}
\details{
Inherits from "VirtualReferenceDataTransformation"
uses "ImpliedFactorReturnsTrendIndicatorTransformationComputation"
}

