% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_transformations.r
\docType{class}
\name{ImpliedFactorReturnsTrendIndicatorTransformationComputation-class}
\alias{ImpliedFactorReturnsTrendIndicatorTransformationComputation-class}
\title{computes trend indicator quantifier}
\description{
Adds columns holding trend indication category for each factor
}
\details{
Inherits from "VirtualTransformationComputation"
}

