% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.r
\docType{methods}
\name{initialize,MarketData-method}
\alias{initialize,MarketData-method}
\title{Initialize method for "MarketData" class}
\usage{
\S4method{initialize}{MarketData}(.Object, symbol, start, end)
}
\arguments{
\item{.Object, }{object of class "MarketData"}

\item{symbol}{"character" name of the symbol queried}

\item{start}{"Date" start date of the data range}

\item{end}{"Date" end date of the data range}
}
\value{
\code{.Object} object of class "MarketData"
}
\description{
Initialize method for "MarketData" class
}

