% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{methods}
\name{initialize,VirtualReferenceDataTransformation-method}
\alias{initialize,VirtualReferenceDataTransformation-method}
\title{Initialize method for "VirtualReferenceDataTransformation" class}
\usage{
\S4method{initialize}{VirtualReferenceDataTransformation}(.Object, ref_data)
}
\arguments{
\item{.Object, }{object of class "VirtualReferenceDataTransformation"}

\item{ref_data}{object of class derived from "VirtualReferenceData"}
}
\value{
\code{.Object} object of class "VirtualReferenceDataTransformation"
}
\description{
Initialize method for "VirtualReferenceDataTransformation" class
}

