% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{methods}
\name{initialize,VirtualRiskModelDataSourceClient-method}
\alias{initialize,VirtualRiskModelDataSourceClient-method}
\title{initialize method for "VirtualRiskModelDataSourceClient" derived classes}
\usage{
\S4method{initialize}{VirtualRiskModelDataSourceClient}(.Object)
}
\arguments{
\item{.Object}{object of class derived from "VirtualRiskModelDataSourceClient"}
}
\description{
initializes required column names from the values obtained from contained risk model
}

