% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_component.r
\docType{methods}
\name{initialize,VirtualRiskModelFactorDependentComponent-method}
\alias{initialize,VirtualRiskModelFactorDependentComponent-method}
\title{initialize method for "VirtualRiskModelFactorDependentComponent" derived classes}
\usage{
\S4method{initialize}{VirtualRiskModelFactorDependentComponent}(.Object)
}
\arguments{
\item{.Object}{object of class derived from "VirtualRiskModelFactorDependentComponent"}
}
\description{
initializes required column names from the values obtained from contained risk model
}

