% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{methods}
\name{initialize,VirtualRiskModelObjectstoreClient-method}
\alias{initialize,VirtualRiskModelObjectstoreClient-method}
\title{initialize method for "VirtualRiskModelObjectstoreClient" derived classes}
\usage{
\S4method{initialize}{VirtualRiskModelObjectstoreClient}(.Object)
}
\arguments{
\item{.Object}{object of class derived from "VirtualRiskModelObjectstoreClient"}
}
\description{
initializes required column names from the values obtained from contained risk model
}

