% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_rodbc_client.r
\docType{methods}
\name{initialize,VirtualRiskModelRODBCClient-method}
\alias{initialize,VirtualRiskModelRODBCClient-method}
\title{initialize method for "VirtualRiskModelRODBCClient" derived classes}
\usage{
\S4method{initialize}{VirtualRiskModelRODBCClient}(.Object)
}
\arguments{
\item{.Object}{object of class derived from "VirtualRiskModelRODBCClient"}
}
\description{
initializes required column names from the values obtained from contained risk model
}

