% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data.r
\docType{class}
\name{InstrumentBetasData-class}
\alias{InstrumentBetasData-class}
\title{Concrtete S4 class handling Instrument Betas data.}
\description{
Implements storage for Instrument Betas and access
to Factor Correlation data via Risk Model Objectstore
}
\details{
Inherits from "VirtualInstrumentBetasData" and "VirtualRiskModelObjectstoreClient"
}

