% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{class}
\name{InvalidRowMeansTransformationComputation-class}
\alias{InvalidRowMeansTransformationComputation-class}
\title{class for test purposes only do not use}
\description{
class for test purposes only do not use
}

