% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{isComputed}
\alias{isComputed}
\alias{isComputed,VirtualBaseTransformation-method}
\title{Is computation done}
\usage{
isComputed(object)

\S4method{isComputed}{VirtualBaseTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{is_computed} logical inicating if computation is done

\code{is_computed} logical inicating if computation is done
}
\description{
Returns logical value indicating if computation has
been done.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualBaseTransformation}: Is computation done

Returns logical value indicating if computation has
been done.
}}

