% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\docType{methods}
\name{isTransformationAttached}
\alias{isTransformationAttached}
\alias{isTransformationAttached,VirtualTransformationsHandler,VirtualReferenceDataTransformation-method}
\title{Check if given transformation is attached}
\usage{
isTransformationAttached(object, transformation)


  \S4method{isTransformationAttached}{VirtualTransformationsHandler,VirtualReferenceDataTransformation}(object,
  transformation)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformation}{object of class "VirtualReferenceDataTransformation"}
}
\value{
\code{is_attached} logical.

\code{is_attached} logical.
}
\description{
Returns logical value indicating if given transformation
is attached to the object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformationsHandler,transformation = VirtualReferenceDataTransformation}: Check if given transformation is attached

Returns logical value indicating if given transformation
is attached to the object
}}

