% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.r
\docType{class}
\name{MarketData-class}
\alias{MarketData-class}
\title{Concrtete S4 class handling Market data from external sources.}
\description{
Implements handling and retrieval of Market data from various sources
via quantmod library. Also computes log returns from Close price.
}
\details{
Inherits from "VirtualDataSourceClient"
}

