% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.r
\docType{class}
\name{MarketDataSX5E-class}
\alias{MarketDataSX5E-class}
\title{Concrtete S4 class handling SX5E index data.}
\description{
Implements handling and retrieval of SX5E index
via quantmod library
}
\details{
Inherits from "MarketData"
}

