% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data.r
\docType{class}
\name{MarketDataSX5P-class}
\alias{MarketDataSX5P-class}
\title{Concrtete S4 class handling SX5P index data.}
\description{
Implements handling and retrieval of SX5P index
via quantmod library
}
\details{
Inherits from "MarketData"
}

