% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data.r
\docType{class}
\name{MarketStyleData-class}
\alias{MarketStyleData-class}
\title{Concrtete S4 class handling Market Style data.}
\description{
Implements storage for Market Style and access
to Market Style data via Risk Model Objectstore
}
\details{
Inherits from "VirtualFactorVarianceData" and "VirtualRiskModelObjectstoreClient"
}

