% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{class}
\name{NullableReferenceData-class}
\alias{NullableReferenceData-class}
\title{NullableReferenceData-class}
\description{
Union class NullableReferenceData
union of "VirtualReferenceData" and "NULL"
}

