% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\docType{class}
\name{Portfolio-class}
\alias{Portfolio-class}
\title{Virtual S4 class implementing handling of portfolio data.}
\description{
Implements handling of portfolio data
handling. Inherits from "VirtualReferenceData".
}
\section{Slots}{

\describe{
\item{\code{start}}{"Date"}

\item{\code{end}}{"Date}
}}

