% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.r
\docType{data}
\name{position_summary_features}
\alias{position_summary_features}
\title{List of position summary features names}
\format{An object of class \code{character} of length 12.}
\usage{
position_summary_features
}
\description{
List of position summary features names
}
\keyword{datasets}

