% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data.r
\docType{class}
\name{PositionData-class}
\alias{PositionData-class}
\title{Concrtete S4 class handling Portfolio Positions data.}
\description{
Implements storage for Portfolio Positions and access
to Portfolio Positions data via warehouse store
}
\details{
Inherits from "VirtualPositionData" and "VirtualWarehouseClient"
}

