% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{.addTransformation}
\alias{.addTransformation}
\title{Add transformation to the object}
\usage{
.addTransformation(object, transformation)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformation}{object of class "VirtualReferenceDataTransformation"}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Private method to add transformation to list of attached transformations
}

