% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.generateDataFilledWithNA}
\alias{.generateDataFilledWithNA}
\title{Fills missing data with NA's}
\usage{
.generateDataFilledWithNA(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{...}{any parameters passed to specific methods implemented by
inheriting classes}
}
\value{
\code{data} "data.frame" with data
}
\description{
Private method to fill data for missing keys with NA's
has to be implemented in derived classes if it is to be used
}

