% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\name{.generateQueryKeyValues}
\alias{.generateQueryKeyValues}
\title{Generate specific objectstore keys}
\usage{
.generateQueryKeyValues(object, key_values)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelObjectstoreClient'.}

\item{key_values}{'data.frame' with original query keys to be transformed .}
}
\value{
\code{query_key_values} 'data.frame', transformed, component specific query keys.
}
\description{
Transforms original query key values to risk model objectstore specific keys
}

