% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLProcedureArgumentNames}
\alias{.getSQLProcedureArgumentNames}
\title{Get stored procedure arg names}
\usage{
.getSQLProcedureArgumentNames(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLProcedureCall'.}
}
\value{
\code{arguments} character vector of names of procedure arguments
}
\description{
Returns names of arguments used by procedure(s).
}

