% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLProcedureName}
\alias{.getSQLProcedureName}
\title{Get stored procedure name}
\usage{
.getSQLProcedureName(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLProcedureCall'.}
}
\value{
\code{procedure} character procedure name
}
\description{
Returns name of the stored procedure that needs to be executed.
}

