% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLQueryKeyValuesParser}
\alias{.getSQLQueryKeyValuesParser}
\title{Get sql query column names}
\usage{
.getSQLQueryKeyValuesParser(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{parser} function used to parse the query keys
}
\description{
Returns function that will be used to parse
query key values to SQL querry string(s).
}

