% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLQueryObject}
\alias{.getSQLQueryObject}
\title{Get SQL query Object}
\usage{
.getSQLQueryObject(object)
}
\arguments{
\item{object}{object of class 'VirtualRODBCClient'.}
}
\value{
\code{sql_query} object of type "VirtualSQLQuery"
}
\description{
Returns stored SQL query object that encapsulates query to DB
}

