% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_rodbc_client.r
\name{.getSQLQueryObjectForRiskComponent}
\alias{.getSQLQueryObjectForRiskComponent}
\title{get specific SQL query for given risk component name}
\usage{
.getSQLQueryObjectForRiskComponent(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{sql_query} object of class "RiskModel.VirtualSQLProcedureCall"
}
\description{
get specific SQL query for given risk component name
}

