% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLQueryResultsParser}
\alias{.getSQLQueryResultsParser}
\title{Get sql query column names}
\usage{
.getSQLQueryResultsParser(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{results_parser} function used to parse the data
}
\description{
Returns function that will be used to parse results of SQL query.
}

