% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLQuerySchemaName}
\alias{.getSQLQuerySchemaName}
\title{Get schema name}
\usage{
.getSQLQuerySchemaName(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{db_schema} character
}
\description{
Returns character name of dB schema to be used for queries.
}

