% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.getSQLQueryStrings}
\alias{.getSQLQueryStrings}
\title{Get sql queries}
\usage{
.getSQLQueryStrings(object)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{sql_query} character vector of constructed SQL query strings.
}
\description{
Returns character vector of constructed SQL query strings.
}

