% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values.r
\name{.obtainValues}
\alias{.obtainValues}
\title{Provides the implementation that defines the
available values.}
\usage{
.obtainValues(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualAvailableValues'.}

\item{...}{any parameters passed to specific methods implemented by
inheriting classes}
}
\value{
\code{values} "data.frame" with values
}
\description{
Should be called within the dataRequest method.
}

