% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.PostProcessResultsData}
\alias{.PostProcessResultsData}
\title{Runs necessary postprocessing routines}
\usage{
.PostProcessResultsData(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Private method to handle cleaning of raw data from datasource
}

