% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.removeNAReferenceData}
\alias{.removeNAReferenceData}
\title{Remove rows with na's}
\usage{
.removeNAReferenceData(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Private method to remove rows in reference data
where non-NA column names are NA
}

