% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{.setAttachedTransformations}
\alias{.setAttachedTransformations}
\title{Set transformation list}
\usage{
.setAttachedTransformations(object, transformations)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformations}{character vector of list of attached transformations}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Private method to set transformation list of attached transformations
}

