% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values_data_handler.r
\name{.setAvailableValuesObject}
\alias{.setAvailableValuesObject}
\title{Set available_Values object in object slot}
\usage{
.setAvailableValuesObject(object, available_Values)
}
\arguments{
\item{object}{object of class "VirtualAvailableValuesDataHandler"}

\item{available_Values}{object of class "VirtualAvailableValues"}
}
\value{
\code{object} object of class "VirtualAvailableValuesDataHandler"
}
\description{
Private method to set available_Values slot with "VirtualAvailableValues"
}

