% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{.setComputation}
\alias{.setComputation}
\title{Set computation object}
\usage{
.setComputation(object, computation)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}

\item{computation}{object of type VirtualTransformationComputation}
}
\value{
\code{compute} function used for computation of specific transformation
}
\description{
Private method to set computation object encapsulated in Transformation .
}

