% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.setDataSourceQueryKeyValues}
\alias{.setDataSourceQueryKeyValues}
\title{Store values of query keys}
\usage{
.setDataSourceQueryKeyValues(object, key_values)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{key_values}{data.frame with columns matching key column names()}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Private method to set values of key values datatable
with keys that will be matched in datastore query
}

