% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.setDataSourceReturnColumnNames}
\alias{.setDataSourceReturnColumnNames}
\title{set column names returned by data source}
\usage{
.setDataSourceReturnColumnNames(object, values)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{values}{"character" vector with list of returned source column names}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Private method to set column names returned by datasource
}

