% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trade_data_handler.r
\name{.setExtendedTradeDataObject}
\alias{.setExtendedTradeDataObject}
\title{Set ex_trade data object in object slot}
\usage{
.setExtendedTradeDataObject(object, ex_trade_data)
}
\arguments{
\item{object}{object of class "VirtualExtendedTradeDataHandler"}

\item{ex_trade_data}{object of class "VirtualExtendedTradeData"}
}
\value{
\code{object} object of class "VirtualExtendedTradeDataHandler"
}
\description{
Private method to set ex_trade_data slot with "VirtualExtendedTradeData"
}

