% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_exposure_data_handler.r
\name{.setFactorExposureDataObject}
\alias{.setFactorExposureDataObject}
\title{Set event data object in object slot}
\usage{
.setFactorExposureDataObject(object, factor_exposure_data)
}
\arguments{
\item{object}{object of class "VirtualFactorExposureDataHandler"}

\item{factor_exposure_data}{object of class "VirtualFactorExposureData"}
}
\value{
\code{object} object of class "VirtualFactorExposureDataHandler"
}
\description{
Private method to set factor_exposure_data slot with "VirtualFactorExposureData"
}

