% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data_handler.r
\name{.setFactorVarianceDataObject}
\alias{.setFactorVarianceDataObject}
\title{Set factor_variance object in object slot}
\usage{
.setFactorVarianceDataObject(object, factor_variance)
}
\arguments{
\item{object}{object of class "VirtualFactorVarianceDataHandler"}

\item{factor_variance}{object of class "VirtualFactorVarianceData"}
}
\value{
\code{object} object of class "VirtualFactorVarianceDataHandler"
}
\description{
Private method to set factor_variance slot with "VirtualFactorVarianceData"
}

