% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data_handler.r
\name{.setImpliedFactorReturnsDataObject}
\alias{.setImpliedFactorReturnsDataObject}
\title{Set implied_factor_returns object in object slot}
\usage{
.setImpliedFactorReturnsDataObject(object, implied_factor_returns)
}
\arguments{
\item{object}{object of class "VirtualImpliedFactorReturnsDataHandler"}

\item{implied_factor_returns}{object of class "VirtualImpliedFactorReturnsData"}
}
\value{
\code{object} object of class "VirtualImpliedFactorReturnsDataHandler"
}
\description{
Private method to set implied_factor_returns slot with "VirtualImpliedFactorReturnsData"
}

