% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_residual_returns_data_handler.r
\name{.setInstrumentResidualReturnsDataObject}
\alias{.setInstrumentResidualReturnsDataObject}
\title{Set instrument_residual_returns object in object slot}
\usage{
.setInstrumentResidualReturnsDataObject(object, instrument_residual_returns)
}
\arguments{
\item{object}{object of class "VirtualInstrumentResidualReturnsDataHandler"}

\item{instrument_residual_returns}{object of class "VirtualInstrumentResidualReturnsData"}
}
\value{
\code{object} object of class "VirtualInstrumentResidualReturnsDataHandler"
}
\description{
Private method to set instrument_residual_returns slot with "VirtualInstrumentResidualReturnsData"
}

