% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{.setIsComputed}
\alias{.setIsComputed}
\title{Set value of computation status}
\usage{
.setIsComputed(object, state = FALSE)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}

\item{state}{logical value to set.}
}
\value{
\code{object} object of class 'VirtualBaseTransformation'
}
\description{
Set value of computation status
}

