% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{.setOutputData}
\alias{.setOutputData}
\title{Set value of computation output}
\usage{
.setOutputData(object, data)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}

\item{data}{data to store in.}
}
\value{
\code{object} object of class 'VirtualTransformationComputation'
}
\description{
Private method to set internal output datastore to new data
}

