% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceobject.r
\name{.setRequiredVariablesNames}
\alias{.setRequiredVariablesNames}
\title{Private function to set required variable names only to be used by derived classes}
\usage{
.setRequiredVariablesNames(object, names)
}
\arguments{
\item{object}{object extending VirtualReferenceObject}

\item{names}{"character" vector with names of required variables}
}
\value{
\code{object} object extending VirtualReferenceObject
}
\description{
Private function to set required variable names only to be used by derived classes
}

