% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.setSQLQueryKeyValues}
\alias{.setSQLQueryKeyValues}
\title{set sql query key values}
\usage{
.setSQLQueryKeyValues(object, key_values)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}

\item{key_values}{data.frame with columns matching key column names()}
}
\value{
\code{object} object of class 'VirtualSQLQuery'.
}
\description{
Private method to set values of key values datatable
with keys that will be matched in datastore query
}

