% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.setSQLQueryObject}
\alias{.setSQLQueryObject}
\title{Set SQL query Object}
\usage{
.setSQLQueryObject(object, sql_query)
}
\arguments{
\item{object}{object of class 'VirtualRODBCClient'.}

\item{sql_query}{bject of type "VirtualSQLQuery"}
}
\value{
\code{object} object of class 'VirtualRODBCClient'.
}
\description{
Sets stored SQL query object that encapsulates query to DB
}

