% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\name{.setSQLQueryStrings}
\alias{.setSQLQueryStrings}
\title{Set sql query strings}
\usage{
.setSQLQueryStrings(object, queries)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}

\item{queries}{character vector of constructed SQL query strings.}
}
\value{
\code{object} object of class 'VirtualSQLQuery'.
}
\description{
Private method to set character vector of constructed SQL query strings.
}

