% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{.setStoredNRows}
\alias{.setStoredNRows}
\title{Set counter of rows}
\usage{
.setStoredNRows(object, nrows)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{nrows}{integer number of stored rows.}
}
\value{
\code{object} object of class 'VirtualReferenceData'.
}
\description{
Private method to set Number of rows of stored data in given ReferenceData class instance.
}

