% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{.setStoredVariablesNames}
\alias{.setStoredVariablesNames}
\title{Set stored variable names}
\usage{
.setStoredVariablesNames(object, names)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{names}{character vector of new Stored Variable names.}
}
\value{
\code{object} object of class 'VirtualReferenceData'.
}
\description{
Private method to set stored variable names only to be used by derived classes
}

