% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{.setTransformations}
\alias{.setTransformations}
\title{Set transformation list}
\usage{
.setTransformations(object, transformations)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformations}{list of objects of class "VirtualReferenceDataTransformation"}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Private method to set transformation list of attached transformations
}

