% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.transformReferenceData}
\alias{.transformReferenceData}
\title{Perform factorization transformation}
\usage{
.transformReferenceData(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Private method to perform unstacking of factor columns
Expands internal factor columns to logical columns
with factor levels as column names
}

