% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.translateDataSourceColumnNames}
\alias{.translateDataSourceColumnNames}
\title{Translate datasource column names}
\usage{
.translateDataSourceColumnNames(object, colnames)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{colnames}{character vector with list of raw data column names}
}
\value{
\code{ret_colnames} data source column names translated to
output column names as required by "required Variables"
}
\description{
Private method to translate raw data column names
to required output names
}

