% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{.updateStoredVariablesNames}
\alias{.updateStoredVariablesNames}
\title{updates stored variable names}
\usage{
.updateStoredVariablesNames(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{object} object of class 'VirtualReferenceData'.
}
\description{
Private method to update variable names only to be used by derived classes
}

