% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{.validateData}
\alias{.validateData}
\title{Validate data against client validation object}
\usage{
.validateData(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{object} client with validated reference data.
}
\description{
Private method to apply the objects validation object against data.
Typically would call this as part of the dataRequest implementation
}

