% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{class}
\name{ReferenceDataTransformationData-class}
\alias{ReferenceDataTransformationData-class}
\title{ReferenceDataTransformationData-class}
\description{
Union class ReferenceDataTransformationData
union of "VirtualReferenceData" and "NULL"
}

