% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_rodbc_client.r
\docType{data}
\name{risk_model_rodbc_keys}
\alias{risk_model_rodbc_keys}
\title{list of key column names for RODBC riskmodel client}
\format{An object of class \code{character} of length 1.}
\usage{
risk_model_rodbc_keys
}
\description{
list of key column names for RODBC riskmodel client
}
\keyword{datasets}

