% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\docType{class}
\name{RiskModel.DevelopedEurope-class}
\alias{RiskModel.DevelopedEurope-class}
\title{Virtual S4 class for Developed Europe Risk Models.}
\description{
Class implementing parameters for family of Developed
Europe Risk Models
}
\details{
Inherits from "VirtualRiskModel"
}

