% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_rodbc_client.r
\docType{class}
\name{RiskModel.VirtualSQLProcedureCall-class}
\alias{RiskModel.VirtualSQLProcedureCall-class}
\title{Virtual S4 class handling sql procedures calls to RiskModel DB.}
\description{
Implements handling of access to data via sql stored procedures calls.
Inherits from "VirtualSQLProcedureCall"
}

