% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{setComputation}
\alias{setComputation}
\title{Set computation object}
\usage{
setComputation(object, computation, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{computation}{object of type VirtualTransformationComputation}
}
\value{
\code{object} object of class 'VirtualTransformation'
}
\description{
sets computation object encapsulated in Transformation.
Handles all checks and settings.
}

