% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{setComputationInput,VirtualTransformation,data.frame-method}
\alias{setComputationInput,VirtualTransformation,data.frame-method}
\title{Set value of computation input}
\usage{
\S4method{setComputationInput}{VirtualTransformation,data.frame}(object, data)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{data}{object of class 'data.frame'.}
}
\value{
\code{object} object of class 'VirtualTransformation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}

