% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{setComputationInput}
\alias{setComputationInput}
\title{Set value of computation input}
\usage{
setComputationInput(object, data, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{data}{object of class 'VirtualReferenceData'.}
}
\value{
\code{object} object of class 'VirtualTransformation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}

