% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{setDataSourceQueryKeyValues}
\alias{setDataSourceQueryKeyValues}
\alias{setDataSourceQueryKeyValues,VirtualDataSourceClient,data.frame-method}
\title{Store values of query keys}
\usage{
setDataSourceQueryKeyValues(object, key_values)


  \S4method{setDataSourceQueryKeyValues}{VirtualDataSourceClient,data.frame}(object,
  key_values)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}

\item{key_values}{data.frame with columns matching key column names()}
}
\value{
\code{object} object of class 'VirtualDataSourceClient'.
}
\description{
Public method to set values of key values datatable
needs to be implemented by inheriting classes
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualDataSourceClient,key_values = data.frame}: Store values of query keys

Public method to set values of key values datatable
}}

