% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\name{setEndDate}
\alias{setEndDate}
\title{Set end date of portfolio timespan}
\usage{
setEndDate(object, end, ...)
}
\arguments{
\item{object}{object of class 'Portfolio'.}

\item{end}{Date end of data range.}
}
\value{
\code{object} object of class 'Portfolio'.
}
\description{
Set end date of portfolio timespan
}

