% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_data_handler.r
\name{setEventDataObject}
\alias{setEventDataObject}
\title{Set event_data object in object slot}
\usage{
setEventDataObject(object, event_data)
}
\arguments{
\item{object}{object of class "VirtualEventDataHandler"}

\item{event_data}{object of class "VirtualEventData"}
}
\value{
\code{object} object of class "VirtualEventDataHandler"
}
\description{
Public method to set event_data slot with "VirtualEventData"
class object to be implemented in derived
}

