% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data_handler.r
\name{setFactorCorrelationDataObject}
\alias{setFactorCorrelationDataObject}
\title{Set factor_correlation data object in object slot}
\usage{
setFactorCorrelationDataObject(object, factor_correlation)
}
\arguments{
\item{object}{object of class "VirtualFactorCorrelationDataHandler"}

\item{factor_correlation}{object of class "VirtualFactorCorrelationData"}
}
\description{
Public method to set factor_correlation slot with "VirtualFactorCorrelationData"
class object to be implemented in derived
}

