% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{setInputData}
\alias{setInputData}
\alias{setInputData,VirtualTransformationComputation,data.frame-method}
\title{Set value of computation input}
\usage{
setInputData(object, data)

\S4method{setInputData}{VirtualTransformationComputation,data.frame}(object,
  data)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}

\item{data}{data to store in.}
}
\value{
\code{object} object of class 'VirtualTransformationComputation'

\code{object} object of class 'VirtualTransformationComputation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformationComputation,data = data.frame}: Set value of computation input

Sets internal input datastore to new data and performs data validity checks .
}}

