% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data_handler.r
\name{setInstrumentBetasDataObject}
\alias{setInstrumentBetasDataObject}
\title{Set instrument_betas object in object slot}
\usage{
setInstrumentBetasDataObject(object, instrument_betas)
}
\arguments{
\item{object}{object of class "VirtualInstrumentBetasDataHandler"}

\item{instrument_betas}{object of class "VirtualInstrumentBetasData"}
}
\value{
\code{object} object of class "VirtualInstrumentBetasDataHandler"
}
\description{
Public method to set instrument_betas slot with "VirtualInstrumentBetasData"
class object to be implemented in derived
}

