% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data_handler.r
\name{setMarketDataObject}
\alias{setMarketDataObject}
\title{Set market_data object in object slot}
\usage{
setMarketDataObject(object, market_data)
}
\arguments{
\item{object}{object of class "VirtualMarketDataHandler"}

\item{market_data}{object of class "VirtualMarketData"}
}
\value{
\code{object} object of class "VirtualMarketDataHandler"
}
\description{
Public method to set market_data slot with "VirtualMarketData"
class object to be implemented in derived
}

