% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data_handler.r
\name{setMarketStyleDataObject}
\alias{setMarketStyleDataObject}
\title{Set market_style object in object slot}
\usage{
setMarketStyleDataObject(object, market_style)
}
\arguments{
\item{object}{object of class "VirtualMarketStyleDataHandler"}

\item{market_style}{object of class "VirtualMarketStyleData"}
}
\value{
\code{object} object of class "VirtualMarketStyleDataHandler"
}
\description{
Public method to set market_style slot with "VirtualMarketStyleData"
class object to be implemented in derived
}

