% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_data_handler.r
\name{setPortfolioDataObject}
\alias{setPortfolioDataObject}
\title{Set portfolio object in object slot}
\usage{
setPortfolioDataObject(object, portfolio, ...)
}
\arguments{
\item{object}{object of class "VirtualPortfolioDataHandler"}

\item{portfolio}{object of class "Portfolio"}
}
\value{
\code{object} object of class "VirtualPortfolioDataHandler"
}
\description{
Public method to set portfolio slot with "Portfolio"
class object to be implemented in derived
}

