% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data_handler.r
\name{setPositionDataObject}
\alias{setPositionDataObject}
\title{Set position_data object in object slot}
\usage{
setPositionDataObject(object, position_data, ...)
}
\arguments{
\item{object}{object of class "VirtualPositionDataHandler"}

\item{position_data}{object of class "VirtualPositionData"}
}
\value{
\code{object} object of class "VirtualPositionDataHandler"
}
\description{
Public method to set position_data slot with "VirtualPositionData"
class object to be implemented in derived
}

