% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data_handler.r
\name{setPriceDataObject}
\alias{setPriceDataObject}
\title{Set price_data object in object slot}
\usage{
setPriceDataObject(object, position_data, ...)
}
\arguments{
\item{object}{object of class "VirtualPriceDataHandler"}

\item{price_data}{object of class "VirtualPriceData"}
}
\value{
\code{object} object of class "VirtualPriceDataHandler"
}
\description{
Public method to set price_data slot with "VirtualPriceData"
class object to be implemented in derived
}

