% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.R
\name{setReferenceData}
\alias{setReferenceData}
\title{Set stored data to value provided}
\usage{
setReferenceData(object, data, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data has to comply with required variable names.}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Sets new values to stored data. Erases any previous data without warning
}

