% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{methods}
\name{setReferenceDataObject}
\alias{setReferenceDataObject}
\alias{setReferenceDataObject,VirtualReferenceDataTransformation,VirtualReferenceData-method}
\title{Set VirtualReferenceData Object}
\usage{
setReferenceDataObject(object, ref_data)


  \S4method{setReferenceDataObject}{VirtualReferenceDataTransformation,VirtualReferenceData}(object,
  ref_data)
}
\arguments{
\item{object}{object of class 'VirtualReferenceDataTransformation'.}

\item{ref_data}{object of class 'VirtualReferenceData'.}
}
\value{
\code{ref_data} object of class 'VirtualReferenceDataTransformation'.

\code{ref_data} object of class 'VirtualReferenceDataTransformation'.
}
\description{
Sets VirtualReferenceData object on which transformation is beeing done
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualReferenceDataTransformation,ref_data = VirtualReferenceData}: Set VirtualReferenceData Object

Sets VirtualReferenceData object on which transformation is beeing done
}}

