% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_component.r
\docType{methods}
\name{setRiskModelObject,VirtualRiskModelFactorDependentComponent,VirtualRiskModel-method}
\alias{setRiskModelObject,VirtualRiskModelFactorDependentComponent,VirtualRiskModel-method}
\title{Set risk_model object in object slot}
\usage{

  \S4method{setRiskModelObject}{VirtualRiskModelFactorDependentComponent,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualRiskModelFactorDependentComponent"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{object} object of class "VirtualRiskModelFactorDependentComponent"
}
\description{
Public method to set trade_data slot with "VirtualRiskModel"
class object
}

