% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\docType{methods}
\name{setStartDate}
\alias{setStartDate}
\alias{setStartDate,Portfolio,Date-method}
\title{Set start date of portfolio timespan}
\usage{
setStartDate(object, start)

\S4method{setStartDate}{Portfolio,Date}(object, start)
}
\arguments{
\item{object}{object of class 'Portfolio'.}

\item{start}{Date start of data range.}
}
\value{
\code{object} object of class 'Portfolio'.

\code{object} object of class 'Portfolio'.
}
\description{
Set start date of portfolio timespan
}
\section{Methods (by class)}{
\itemize{
\item \code{object = Portfolio,start = Date}: Set start date of portfolio timespan
}}

