% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_data_handler.r
\name{setStrategyDataObject}
\alias{setStrategyDataObject}
\title{Set strategy_data object in object slot}
\usage{
setStrategyDataObject(object, strategy_data, ...)
}
\arguments{
\item{object}{object of class "VirtualStrategyDataHandler"}

\item{strategy_data}{object of class "VirtualStrategyData"}
}
\value{
\code{object} object of class "VirtualStrategyDataHandler"
}
\description{
Public method to set strategy_data slot with "VirtualStrategyData"
class object to be implemented in derived
}

