% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data_handler.r
\name{setTradeDataObject}
\alias{setTradeDataObject}
\title{Set trade_data object in object slot}
\usage{
setTradeDataObject(object, trade_data, ...)
}
\arguments{
\item{object}{object of class "VirtualTradeDataHandler"}

\item{trade_data}{object of class "VirtualTradeData"}
}
\value{
\code{object} object of class "VirtualTradeDataHandler"
}
\description{
Public method to set trade_data slot with "VirtualTradeData"
class object to be implemented in derived
}

