% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_portfolio.r
\docType{class}
\name{StrategyPortfolio-class}
\alias{StrategyPortfolio-class}
\title{S4 class implementing handling of strategy portfolio}
\description{
This is concrete class handling data retrieval
of strategy portfolio data.
Inherits from:
 "Portfolio"
 "VirtualTransformationsHandler"
 "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{trader_id}}{"numeric"}
}}

