% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TE.RefClasses.r
\docType{package}
\name{TE.RefClasses}
\alias{TE.RefClasses}
\alias{TE.RefClasses-package}
\title{TE.RefClasses: Trading Enhancement Data access, cashing and preprocessing}
\description{
The TE.RefClasses package providing classes and functions for data retrieval,
caching and manipulation for Trading Enhancement R engine via reference data
builds on top of TE.DataAccess package
}

