% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client_test_class.r
\docType{class}
\name{TestDataStoreClient-class}
\alias{TestDataStoreClient-class}
\title{Concrete S4 class for testing of DataStoreClient.}
\description{
Test class for testing VirtualDataStoreClient
functionality. Do not use.
}
\details{
Inherits from "VirtualDataStoreClient"
}

