% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.r
\docType{data}
\name{trade_basic_features}
\alias{trade_basic_features}
\title{List of trade history basic features names}
\format{An object of class \code{character} of length 12.}
\usage{
trade_basic_features
}
\description{
List of trade history basic features names
}
\keyword{datasets}

