% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.r
\docType{data}
\name{trade_basic_features_transl}
\alias{trade_basic_features_transl}
\title{List of translated trade history basic features names}
\format{An object of class \code{character} of length 11.}
\usage{
trade_basic_features_transl
}
\description{
List of translated trade history basic features names
}
\keyword{datasets}

