% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.r
\docType{data}
\name{trade_delta_features}
\alias{trade_delta_features}
\title{List of trade delta features names}
\format{An object of class \code{character} of length 3.}
\usage{
trade_delta_features
}
\description{
List of trade delta features names
}
\keyword{datasets}

