% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data.r
\docType{data}
\name{trade_history_features}
\alias{trade_history_features}
\title{List of trade history features names}
\format{An object of class \code{character} of length 2.}
\usage{
trade_history_features
}
\description{
List of trade history features names
}
\keyword{datasets}

