% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data_simple.r
\docType{class}
\name{TradeDataSimple-class}
\alias{TradeDataSimple-class}
\title{Concrete S4 class handling Simple TradeData}
\description{
Inherits from "VirtualTradeData" and "VirtualPPModelClient"
}

