% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data_handler.r
\docType{methods}
\name{updateFactorCorrelationDataRiskModel}
\alias{updateFactorCorrelationDataRiskModel}
\alias{updateFactorCorrelationDataRiskModel,VirtualFactorCorrelationDataHandler,VirtualRiskModel-method}
\title{Update factor_correlation risk model}
\usage{
updateFactorCorrelationDataRiskModel(object, risk_model)


  \S4method{updateFactorCorrelationDataRiskModel}{VirtualFactorCorrelationDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualFactorCorrelationDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{factor_correlation} object of class "VirtualFactorCorrelationData"
}
\description{
Trigger update of risk model of factor_correlation object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualFactorCorrelationDataHandler,risk_model = VirtualRiskModel}: Update factor_correlation risk model

Trigger update of risk model of factor_correlation object
}}

