% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data_handler.r
\docType{methods}
\name{updateFactorVarianceDataRiskModel}
\alias{updateFactorVarianceDataRiskModel}
\alias{updateFactorVarianceDataRiskModel,VirtualFactorVarianceDataHandler,VirtualRiskModel-method}
\title{Update factor_variance risk model}
\usage{
updateFactorVarianceDataRiskModel(object, risk_model)


  \S4method{updateFactorVarianceDataRiskModel}{VirtualFactorVarianceDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualFactorVarianceDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{factor_variance} object of class "VirtualFactorVarianceData"
}
\description{
Trigger update of risk model of factor_variance object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualFactorVarianceDataHandler,risk_model = VirtualRiskModel}: Update factor_variance risk model

Trigger update of risk model of factor_variance object
}}

