% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data_handler.r
\docType{methods}
\name{updateImpliedFactorReturnsDataRiskModel}
\alias{updateImpliedFactorReturnsDataRiskModel}
\alias{updateImpliedFactorReturnsDataRiskModel,VirtualImpliedFactorReturnsDataHandler,VirtualRiskModel-method}
\title{Update implied_factor_returns risk model}
\usage{
updateImpliedFactorReturnsDataRiskModel(object, risk_model)


  \S4method{updateImpliedFactorReturnsDataRiskModel}{VirtualImpliedFactorReturnsDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualImpliedFactorReturnsDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{implied_factor_returns} object of class "VirtualImpliedFactorReturnsData"
}
\description{
Trigger update of risk model of implied_factor_returns object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualImpliedFactorReturnsDataHandler,risk_model = VirtualRiskModel}: Update implied_factor_returns risk model

Trigger update of risk model of implied_factor_returns object
}}

