% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data_handler.r
\docType{methods}
\name{updateInstrumentBetasDataRiskModel}
\alias{updateInstrumentBetasDataRiskModel}
\alias{updateInstrumentBetasDataRiskModel,VirtualInstrumentBetasDataHandler,VirtualRiskModel-method}
\title{Update instrument_betas risk model}
\usage{
updateInstrumentBetasDataRiskModel(object, risk_model)


  \S4method{updateInstrumentBetasDataRiskModel}{VirtualInstrumentBetasDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualInstrumentBetasDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{instrument_betas} object of class "VirtualInstrumentBetasData"
}
\description{
Trigger update of risk model of instrument_betas object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualInstrumentBetasDataHandler,risk_model = VirtualRiskModel}: Update instrument_betas risk model

Trigger update of risk model of instrument_betas object
}}

