% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_residual_returns_data_handler.r
\docType{methods}
\name{updateInstrumentResidualReturnsDataRiskModel}
\alias{updateInstrumentResidualReturnsDataRiskModel}
\alias{updateInstrumentResidualReturnsDataRiskModel,VirtualInstrumentResidualReturnsDataHandler,VirtualRiskModel-method}
\title{Update instrument_residual_returns risk model}
\usage{
updateInstrumentResidualReturnsDataRiskModel(object, risk_model)


  \S4method{updateInstrumentResidualReturnsDataRiskModel}{VirtualInstrumentResidualReturnsDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualInstrumentResidualReturnsDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{instrument_residual_returns} object of class "VirtualInstrumentResidualReturnsData"
}
\description{
Trigger update of risk model of instrument_residual_returns object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualInstrumentResidualReturnsDataHandler,risk_model = VirtualRiskModel}: Update instrument_residual_returns risk model

Trigger update of risk model of instrument_residual_returns object
}}

