% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data_handler.r
\docType{methods}
\name{updateMarketStyleDataRiskModel}
\alias{updateMarketStyleDataRiskModel}
\alias{updateMarketStyleDataRiskModel,VirtualMarketStyleDataHandler,VirtualRiskModel-method}
\title{Update market_style risk model}
\usage{
updateMarketStyleDataRiskModel(object, risk_model)


  \S4method{updateMarketStyleDataRiskModel}{VirtualMarketStyleDataHandler,VirtualRiskModel}(object,
  risk_model)
}
\arguments{
\item{object}{object of class "VirtualMarketStyleDataHandler"}

\item{risk_model}{object of class "VirtualRiskModel"}
}
\value{
\code{market_style} object of class "VirtualMarketStyleData"
}
\description{
Trigger update of risk model of market_style object
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualMarketStyleDataHandler,risk_model = VirtualRiskModel}: Update market_style risk model

Trigger update of risk model of market_style object
}}

