% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{updateVariables}
\alias{updateVariables}
\alias{updateVariables,VirtualReferenceData,data.frame,character-method}
\title{Update values in specified columns}
\usage{
updateVariables(object, data, var.names)

\S4method{updateVariables}{VirtualReferenceData,data.frame,character}(object,
  data, var.names)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data has to comply with required variable names.}

\item{var.names}{column names of data in both incomming data and stored data}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.

\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Sets new values to specified columns in stored data.
The number of rows of incoming data has to be the same as number of
rows currently stored.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualReferenceData,data = data.frame,var.names = character}: Update values in specified columns

Sets new values to specified columns in stored data.
The number of rows of incoming data has to be the same as number of
rows currently stored.
}}

