% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validator.r
\name{validateData}
\alias{validateData}
\title{Apply validation to dataframe}
\usage{
validateData(object, data)
}
\arguments{
\item{object}{object of class 'VirtualDataValidator'.}
}
\value{
\code{data} data.frame of validated data
}
\description{
Public method to return dataframe post validation
or halt execution of halt is TRUE
}

