% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values.r
\docType{class}
\name{VirtualAvailableValues-class}
\alias{VirtualAvailableValues-class}
\title{Virtual S4 class providing a mechanism for
to query the possible values of a variable.}
\description{
.obtainValues implements the method by
which the values are determined.
}
\details{
dataRequest implementation supplies the
values.

Inherits from "VirtualDataSourceClient
}

