% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values_data_handler.r
\docType{class}
\name{VirtualAvailableValuesDataHandler-class}
\alias{VirtualAvailableValuesDataHandler-class}
\title{Virtual S4 class handling VirtualAvailableValues objects}
\description{
Class that is to be inherited by any objects
that will contain AvailableValues
}
\section{Slots}{

\describe{
\item{\code{Values}}{object of class "VirtualAvailableValues"}
}}

