% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{class}
\name{VirtualBaseTransformation-class}
\alias{VirtualBaseTransformation-class}
\title{Virtual S4 class implementing basic transformation}
\description{
This is basic class defining Tranformation base
slots and accessors. Inherits from "VirtualReferenceObject"
}
\section{Slots}{

\describe{
\item{\code{computed_colnms}}{"character"}

\item{\code{is_computed}}{"logical"}
}}

