% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{class}
\name{VirtualDataSourceClient-class}
\alias{VirtualDataSourceClient-class}
\title{Virtual S4 class implementing handling of data access.}
\description{
Implements handling of access to data sources.
Handles queries and implements accessors to
internal data.
Inherits from "VirtualReferenceData"
}
\section{Slots}{

\describe{
\item{\code{key_cols}}{= "character",  query keys column names}

\item{\code{key_values}}{= "data.frame",  query keys}

\item{\code{values}}{= "character",  columns that neeed to be returned from datastore}

\item{\code{column_name_map}}{= "hash",  perl-type hash containing mapping of source column names to output column names}

\item{\code{non_na_cols}}{= "character",  columns upon which we remove rows to forced to have non NA values}

\item{\code{factorized_cols}}{= "character",  factor columns that are expanded to  new columns}

\item{\code{factorization_keys}}{= "character", key columns that are used as a unique keys in factorization}
}}

