% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client.r
\docType{class}
\name{VirtualDataStoreClient-class}
\alias{VirtualDataStoreClient-class}
\title{Virtual S4 class implementing handling of data access.}
\description{
Implements handling of access to data via DataStore class.
Inherits from "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{datastore_name}}{= "character",  name of datastore}
}}

