% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validator.r
\docType{class}
\name{VirtualDataValidator-class}
\alias{VirtualDataValidator-class}
\title{Virtual S4 class defining data validation operations}
\description{
Class that is to be inherited by any objects
that will contain VirtualDataValidator or derived classes
}
\details{
The validator applies a test to a data frame.
The test is in the form of a selection of data within the frame as
specifed by some function.
If the data passes this test then it is validated.
If the size of the data would be modified by this selection then the data
fails validation, in which case the data passing the validation can be returned
and a warning raised, or the validator can halt execution.
To construct the validator there needs to be a function that takes as its arguments,
the dataframe to be validated and an optional set of arguments that each represent tnnamed list having names matching the
}
\section{Slots}{

\describe{
\item{\code{validation_function}}{"function", function applied to the data frame to be validated}

\item{\code{argument_list,}}{named list linking the parameter names of the validation_function to column names}

\item{\code{halt,}}{if true execution is halted on failed validation}

\item{\code{mssg}}{displyed if validation is failed}
}}

