% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_data_handler.r
\docType{class}
\name{VirtualEventDataHandler-class}
\alias{VirtualEventDataHandler-class}
\title{Virtual S4 class handling VirtualEventData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualEventData or derived classes
}
\section{Slots}{

\describe{
\item{\code{event_data}}{object of class "VirtualEventData"}
}}

