% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trade_data_handler.r
\docType{class}
\name{VirtualExtendedTradeDataHandler-class}
\alias{VirtualExtendedTradeDataHandler-class}
\title{Virtual S4 class handling VirtualExtendedTradeData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualExtendedTradeData or derived classes
}
\section{Slots}{

\describe{
\item{\code{ex_trade_data}}{object of class "VirtualExtendedTradeData"}
}}

