% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data.r
\docType{class}
\name{VirtualFactorCorrelationData-class}
\alias{VirtualFactorCorrelationData-class}
\title{Virtual S4 class handling Factor Correlation data.}
\description{
Implements storage for Factor Correlation
}
\details{
Inherits from "VirtualReferenceData
}

