% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data_handler.r
\docType{class}
\name{VirtualFactorCorrelationDataHandler-class}
\alias{VirtualFactorCorrelationDataHandler-class}
\title{Virtual S4 class handling VirtualFactorCorrelationData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualFactorCorrelationData or derived classes
}
\section{Slots}{

\describe{
\item{\code{factor_correlation}}{object of class "VirtualFactorCorrelationData"}
}}

