% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_exposure_data_handler.r
\docType{class}
\name{VirtualFactorExposureDataHandler-class}
\alias{VirtualFactorExposureDataHandler-class}
\title{Virtual S4 class handling VirtualFactorExposureData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualFactorExposureData or derived classes
}
\section{Slots}{

\describe{
\item{\code{factor_exposure_data}}{object of class "VirtualFactorExposureData"}
}}

