% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data.r
\docType{class}
\name{VirtualFactorVarianceData-class}
\alias{VirtualFactorVarianceData-class}
\title{Virtual S4 class handling Factor Variance data.}
\description{
Implements storage for Factor Variance
}
\details{
Inherits from "VirtualReferenceData
}

