% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data_handler.r
\docType{class}
\name{VirtualFactorVarianceDataHandler-class}
\alias{VirtualFactorVarianceDataHandler-class}
\title{Virtual S4 class handling VirtualFactorVarianceData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualFactorVarianceData or derived classes
}
\section{Slots}{

\describe{
\item{\code{factor_variance}}{object of class "VirtualFactorVarianceData"}
}}

