% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data.r
\docType{class}
\name{VirtualImpliedFactorReturnsData-class}
\alias{VirtualImpliedFactorReturnsData-class}
\title{Virtual S4 class handling Implied Factor Returns data.}
\description{
Implements storage for Implied Factor Returns
}
\details{
Inherits from "VirtualReferenceData
}

