% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data_handler.r
\docType{class}
\name{VirtualImpliedFactorReturnsDataHandler-class}
\alias{VirtualImpliedFactorReturnsDataHandler-class}
\title{Virtual S4 class handling VirtualImpliedFactorReturnsData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualImpliedFactorReturnsData or derived classes
}
\section{Slots}{

\describe{
\item{\code{implied_factor_returns}}{object of class "VirtualImpliedFactorReturnsData"}
}}

