% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_portfolio.r
\docType{class}
\name{VirtualIndexPortfolio-class}
\alias{VirtualIndexPortfolio-class}
\title{Virtual S4 class handling IndexPortfolio Data}
\description{
Inherits from
 "Portfolio"
 "VirtualRODBCClient"
}
\section{Slots}{

\describe{
\item{\code{index_ticker}}{"character" name of the index ticker as by BBG}
}}

