% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data_handler.r
\docType{class}
\name{VirtualInstrumentBetasDataHandler-class}
\alias{VirtualInstrumentBetasDataHandler-class}
\title{Virtual S4 class handling VirtualInstrumentBetasData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualInstrumentBetasData or derived classes
}
\section{Slots}{

\describe{
\item{\code{instrument_betas}}{object of class "VirtualInstrumentBetasData"}
}}

