% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_residual_returns_data.r
\docType{class}
\name{VirtualInstrumentResidualReturnsData-class}
\alias{VirtualInstrumentResidualReturnsData-class}
\title{Virtual S4 class handling Instrument Betas data.}
\description{
Implements storage for Factor Correlation
}
\details{
Inherits from "VirtualReferenceData"
}

