% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_residual_returns_data_handler.r
\docType{class}
\name{VirtualInstrumentResidualReturnsDataHandler-class}
\alias{VirtualInstrumentResidualReturnsDataHandler-class}
\title{Virtual S4 class handling VirtualInstrumentResidualReturnsData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualInstrumentResidualReturnsData or derived classes
}
\section{Slots}{

\describe{
\item{\code{instrument_residual_returns}}{object of class "VirtualInstrumentResidualReturnsData"}
}}

