% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_expander.r
\docType{class}
\name{VirtualKeyExpander-class}
\alias{VirtualKeyExpander-class}
\title{Virtual S4 class that expands a key column over
the available values from the relevant handler.}
\description{
Virtual S4 class that expands a key column over
the available values from the relevant handler.
}
\section{Slots}{

\describe{
\item{\code{traders}}{object of class "VirtualAvailableValues"}
}}

