% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data_handler.r
\docType{class}
\name{VirtualMarketDataHandler-class}
\alias{VirtualMarketDataHandler-class}
\title{Virtual S4 class handling VirtualMarketData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualMarketData or derived classes
}
\section{Slots}{

\describe{
\item{\code{market_data}}{object of class "VirtualMarketData"}
}}

