% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data.r
\docType{class}
\name{VirtualMarketStyleData-class}
\alias{VirtualMarketStyleData-class}
\title{Virtual S4 class handling Market Style data.}
\description{
Implements storage for Market Style
}
\details{
Inherits from "VirtualReferenceData
}

