% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data_handler.r
\docType{class}
\name{VirtualMarketStyleDataHandler-class}
\alias{VirtualMarketStyleDataHandler-class}
\title{Virtual S4 class handling VirtualMarketStyleData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualMarketStyleData or derived classes
}
\section{Slots}{

\describe{
\item{\code{market_style}}{object of class "VirtualMarketStyleData"}
}}

