% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_data_handler.r
\docType{class}
\name{VirtualPortfolioDataHandler-class}
\alias{VirtualPortfolioDataHandler-class}
\title{Virtual S4 class handling Portfolio objects}
\description{
Class that is to be inherited by any objects
that will contain Portfolio or derived classes
}
\section{Slots}{

\describe{
\item{\code{portfolio}}{object of class "Portfolio"}
}}

