% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data.r
\docType{class}
\name{VirtualPositionData-class}
\alias{VirtualPositionData-class}
\title{Virtual S4 class handling Portfolio Positions data.}
\description{
Implements storage Portfolio Positions
}
\details{
Inherits from "VirtualReferenceData
}

