% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data_handler.r
\docType{class}
\name{VirtualPositionDataHandler-class}
\alias{VirtualPositionDataHandler-class}
\title{Virtual S4 class handling VirtualPositionData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualPositionData or derived classes
}
\section{Slots}{

\describe{
\item{\code{position_data}}{object of class "VirtualPositionData"}
}}

