% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmodel_client.r
\docType{class}
\name{VirtualPPModelClient-class}
\alias{VirtualPPModelClient-class}
\title{Virtual S4 class handling PPModel Objectstore access}
\description{
Implements storage Portfolio Positions
}
\details{
Inherits from "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{model_class}}{"character"}
}}

