% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data_handler.r
\docType{class}
\name{VirtualPriceDataHandler-class}
\alias{VirtualPriceDataHandler-class}
\title{Virtual S4 class handling VirtualPriceData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualPriceData or derived classes
}
\section{Slots}{

\describe{
\item{\code{price_data}}{object of class "VirtualPriceData"}
}}

