% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.R
\docType{class}
\name{VirtualReferenceData-class}
\alias{VirtualReferenceData-class}
\title{Virtual S4 class implementing handling of reference data.}
\description{
Implements handling of required variables and (tbd) error
handling. Inherits from "VirtualReferenceObject".
}
\section{Slots}{

\describe{
\item{\code{column_names}}{"character"}

\item{\code{data}}{"data.frame"}

\item{\code{unique_rows}}{"logical"}

\item{\code{stored_rows}}{"numeric"}
}}

