% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{class}
\name{VirtualReferenceDataTransformation-class}
\alias{VirtualReferenceDataTransformation-class}
\title{Virtual S4 class implementing reference data transformations}
\description{
This is base class handling Tranformation
on VirtualReferenceData object
}
\section{Slots}{

\describe{
\item{\code{ref_data}}{"VirtualReferenceData"}
}}

