% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceobject.r
\docType{class}
\name{VirtualReferenceObject-class}
\alias{VirtualReferenceObject-class}
\title{Virtual S4 class implementing handling of required variables.}
\description{
Implements handling of required variables and (tbd) error
handling
}
\section{Slots}{

\describe{
\item{\code{has_error}}{"logical"}

\item{\code{error_list}}{"character"}

\item{\code{last_error}}{"character"}

\item{\code{required_colnms}}{"character"}
}}

