% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_rodbc_client.r
\docType{class}
\name{VirtualRiskModelClientPicker-class}
\alias{VirtualRiskModelClientPicker-class}
\title{Virtual S4 class for access to Risk Model Objectstore.}
\description{
This is a class serving as a switcher between RODBC and Objectstore
Source of data for risk model
}
\details{
Inherits from "VirtualDataSourceClient" and "VirtualRiskModelHandler"
}
\section{Slots}{

\describe{
\item{\code{component}}{"character", name of the component of risk model}
}}

