% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{class}
\name{VirtualRiskModelDataSourceClient-class}
\alias{VirtualRiskModelDataSourceClient-class}
\title{Virtual S4 class for access to Risk Model Components.}
\description{
This is handler class that is to be inherited
by other classes handling Risk Model Objects
}
\details{
Inherits from "VirtualDataSourceClient"
}
\section{Slots}{

\describe{
\item{\code{component}}{"character", name of the component of risk model}
}}

