% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\docType{class}
\name{VirtualRiskModelObjectstoreClient-class}
\alias{VirtualRiskModelObjectstoreClient-class}
\title{Virtual S4 class for access to Risk Model Objectstore.}
\description{
This is handler class that is to be inherited
by other classes handling Risk Model Objects
}
\details{
Inherits from "VirtualDataSourceClient" and "VirtualRiskModelHandler"
}
\section{Slots}{

\describe{
\item{\code{component}}{"character", name of the component of risk model}
}}

