% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\docType{class}
\name{VirtualRODBCClient-class}
\alias{VirtualRODBCClient-class}
\title{Virtual S4 class wrapping VirtualSQLQuery.}
\description{
Implements handling of access to data via sql stored procedures calls.
Provides SQL data access interface to Reference Classes
}
\section{Slots}{

\describe{
\item{\code{sql_query}}{= "VirtualRODBCClient",}
}}

