% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\docType{class}
\name{VirtualSQLQuery-class}
\alias{VirtualSQLQuery-class}
\title{Virtual S4 class sql queries handling.}
\description{
Implements handling of access to data via SQL queries.
}
\section{Slots}{

\describe{
\item{\code{db_schema}}{"character",}

\item{\code{sql_query}}{"character",}

\item{\code{key_cols}}{"character",}

\item{\code{key_values}}{"data.frame",}

\item{\code{query_parser}}{"function",}

\item{\code{results_parser}}{"function"}
}}

