% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_data_handler.r
\docType{class}
\name{VirtualStrategyDataHandler-class}
\alias{VirtualStrategyDataHandler-class}
\title{Virtual S4 class handling VirtualStrategyData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualStrategyData or derived classes
}
\section{Slots}{

\describe{
\item{\code{strategy_data}}{object of class "VirtualStrategyData"}
}}

