% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data_handler.r
\docType{class}
\name{VirtualTradeDataHandler-class}
\alias{VirtualTradeDataHandler-class}
\title{Virtual S4 class handling VirtualTradeData objects}
\description{
Class that is to be inherited by any objects
that will contain VirtualTradeData or derived classes
}
\section{Slots}{

\describe{
\item{\code{trade_data}}{object of class "VirtualTradeData"}
}}

