% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traderid_key_expander.r
\docType{class}
\name{VirtualTradeIDKeyExpander-class}
\alias{VirtualTradeIDKeyExpander-class}
\title{Virtual S4 class that expands a key frame over
available values of TradeID}
\description{
Virtual S4 class that expands a key frame over
available values of TradeID
}
\section{Slots}{

\describe{
\item{\code{available_values}}{object of class "AvailableTraders"}
}}

