% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{class}
\name{VirtualTransformation-class}
\alias{VirtualTransformation-class}
\title{Virtual S4 class implementing transformations}
\description{
This is adaptation class wrapping computation
class  and preparing computation input and
output for calling entity. Inherits
from "VirtualBaseTransformation"
}
\section{Slots}{

\describe{
\item{\code{computation}}{"VirtualTransformationComputation"}
}}

