% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{class}
\name{VirtualTransformationComputation-class}
\alias{VirtualTransformationComputation-class}
\title{Virtual S4 class implementing basic transformation}
\description{
This is wrapper class wrapping computation
function and managing validation of variables
for input and output. Inherits from "VirtualBaseTransformation"
}
\section{Slots}{

\describe{
\item{\code{input}}{"character"}

\item{\code{output}}{"logical"}

\item{\code{compute}}{"function"}
}}

