% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\docType{class}
\name{VirtualTransformationsHandler-class}
\alias{VirtualTransformationsHandler-class}
\title{Virtual S4 class handing transformations}
\description{
implements handling of transformations
}
\section{Slots}{

\describe{
\item{\code{transformations}}{"list"}

\item{\code{attached_transformations}}{"character"}
}}

