% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warehouse_client.r
\docType{class}
\name{VirtualWarehouseClient-class}
\alias{VirtualWarehouseClient-class}
\title{Virtual S4 class handling data access via warehouse.}
\description{
Inherits from "VirtualDataSourceClient
}

