% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.R
\name{appendVariables}
\alias{appendVariables}
\title{Append new columns to stored data}
\usage{
appendVariables(object, data, var.names)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data}

\item{var.names}{column names of data in both incomming data}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Adds new columns to stored data
The number of rows of incoming data has to be the same as number of
rows currently stored.
}

