% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\docType{methods}
\name{attachTransformation}
\alias{attachTransformation}
\alias{attachTransformation,VirtualTransformationsHandler,VirtualReferenceDataTransformation-method}
\title{Attach transformation to the object}
\usage{
attachTransformation(object, transformation, replace)


  \S4method{attachTransformation}{VirtualTransformationsHandler,VirtualReferenceDataTransformation}(object,
  transformation, replace = TRUE)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformation}{objects of clas "VirtualReferenceDataTransformation"}

\item{replace}{"logical" replace transformation if already attached.}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.

\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformationsHandler,transformation = VirtualReferenceDataTransformation}: Attach transformation to the object

Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}}

