% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{attachTransformations}
\alias{attachTransformations}
\title{Attach transformations to the object}
\usage{
attachTransformations(object, transformations, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformations}{list of objects of class "VirtualReferenceDataTransformation"}
}
\value{
\code{object} object of class 'VirtualTransformationHandler'.
}
\description{
Attaches new columns with new transformations
(data computed from existing VirtualTransformationHandler data)
}

