% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key_expander.r
\docType{methods}
\name{expandKeys}
\alias{expandKeys}
\alias{expandKeys,VirtualKeyExpander,data.frame-method}
\title{Expand key frame over the available_values column name
if that column is not already present.}
\usage{
expandKeys(object, key_values)

\S4method{expandKeys}{VirtualKeyExpander,data.frame}(object, key_values)
}
\arguments{
\item{object}{object of class "VirtualKeyExpander"}

\item{object}{key_values of class "data.frame"}
}
\value{
\code{object} object of class "VirtualKeyExpander"

\code{object} object of class "VirtualKeyExpander"
}
\description{
Expand key frame over the available_values column name
if that column is not already present.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualKeyExpander,key_values = data.frame}: Expand key frame over the available_values column name
if that column is not already present.
}}

