% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getComputation}
\alias{getComputation}
\title{Get computation object}
\usage{
getComputation(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{compute} function used for computation of specific transformation
}
\description{
Returns computation object encapsulated in Transformation .
}

