% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getComputationInput}
\alias{getComputationInput}
\title{Get value of computation input}
\usage{
getComputationInput(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}
}
\value{
\code{data} data.frame with transformation computation input
}
\description{
Returns internal input datastore of computation object .
}

