% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getComputationOutput}
\alias{getComputationOutput}
\title{Get value of computation output}
\usage{
getComputationOutput(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}
}
\value{
\code{data} data.frame with transformation computation output
}
\description{
Returns internal output datastore of computation object .
}

