% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getComputedVariablesNames}
\alias{getComputedVariablesNames}
\title{Get names of computed columns}
\usage{
getComputedVariablesNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{computed_colnms} character vector of names of computed columns
}
\description{
Returns Names of the variables that are generated by computation object.
}

