% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{getDataSourceClientColumnNameMap}
\alias{getDataSourceClientColumnNameMap}
\alias{getDataSourceClientColumnNameMap,VirtualDataSourceClient-method}
\title{Retreive datasource column name map}
\usage{
getDataSourceClientColumnNameMap(object)

\S4method{getDataSourceClientColumnNameMap}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{column_name_map} hash column name map of the datasource to output names

\code{column_name_map} hash column name map of the datasource to output names
}
\description{
Returns hash with mapping of datasource column names
to reference data class column names
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Retreive datasource column name map

Returns hash with mapping of datasource column names
to reference data class column names
}}

