% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{getDataSourceQueryKeyColumnNames}
\alias{getDataSourceQueryKeyColumnNames}
\title{Get query key column names}
\usage{
getDataSourceQueryKeyColumnNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{key_cols} character vector with list of key column names
}
\description{
Private method to return key column names
that are used to query raw data source
}

