% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{getDataSourceQueryKeyValues}
\alias{getDataSourceQueryKeyValues}
\title{Retreive values of query keys}
\usage{
getDataSourceQueryKeyValues(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_non_na_cols} logical
}
\description{
Get values of key values datatable
with keys used to  match in datasource query
}

