% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{getDataSourceReturnColumnNames}
\alias{getDataSourceReturnColumnNames}
\alias{getDataSourceReturnColumnNames,VirtualDataSourceClient-method}
\title{Get returned source data column names}
\usage{
getDataSourceReturnColumnNames(object)

\S4method{getDataSourceReturnColumnNames}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{values} character vector with list of returned source column names

\code{values} character vector with list of returned source column names
}
\description{
Returns names of the columns that will be returned from data source
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Get returned source data column names

Returns names of the columns that will be returned from data source
}}

