% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore_client.r
\docType{methods}
\name{getDataStoreName}
\alias{getDataStoreName}
\alias{getDataStoreName,VirtualDataStoreClient-method}
\title{Get source datastore name}
\usage{
getDataStoreName(object)

\S4method{getDataStoreName}{VirtualDataStoreClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataStoreClient'.}
}
\value{
\code{datastore_name} character

\code{datastore_name} character
}
\description{
Returns name of the datastore which will be querried when asked to fill data.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataStoreClient}: Get source datastore name

Returns name of the datastore which will be querried when asked to fill data.
}}

