% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\docType{methods}
\name{getEndDate}
\alias{getEndDate}
\alias{getEndDate,Portfolio-method}
\title{Get end date of portfolio timespan}
\usage{
getEndDate(object)

\S4method{getEndDate}{Portfolio}(object)
}
\arguments{
\item{object}{object of class 'Portfolio'.}
}
\value{
\code{end} Date end of data range.

\code{end} Date end of data range.
}
\description{
Get end date of portfolio timespan
}
\section{Methods (by class)}{
\itemize{
\item \code{Portfolio}: Get end date of portfolio timespan
}}

