% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_data_handler.r
\name{getEventDataObject}
\alias{getEventDataObject}
\title{Get event_data stored in object}
\usage{
getEventDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualEventDataHandler"}
}
\value{
\code{event_data} object of class "VirtualEventData"
}
\description{
Returns event_data object of class "VirtualEventData"
}

