% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_trade_data_handler.r
\name{getExtendedTradeDataObject}
\alias{getExtendedTradeDataObject}
\title{Get ex_trade data stored in object}
\usage{
getExtendedTradeDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualExtendedTradeDataHandler"}
}
\value{
\code{ex_trade_data} object of class "VirtualExtendedTradeData"
}
\description{
Returns ex_trade_data object of class "VirtualExtendedTradeData"
}

