% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_correlation_data_handler.r
\docType{methods}
\name{getFactorCorrelationDataObject}
\alias{getFactorCorrelationDataObject}
\alias{getFactorCorrelationDataObject,VirtualFactorCorrelationDataHandler-method}
\title{Get factor_correlation data stored in object}
\usage{
getFactorCorrelationDataObject(object)


  \S4method{getFactorCorrelationDataObject}{VirtualFactorCorrelationDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualFactorCorrelationDataHandler"}
}
\value{
\code{factor_correlation} object of class "VirtualFactorCorrelationData"

\code{factor_correlation} object of class "VirtualFactorCorrelationData"
}
\description{
Returns factor_correlation object of class "VirtualFactorCorrelationData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualFactorCorrelationDataHandler}: Get factor_correlation data stored in object

Returns factor_correlation object of class "VirtualFactorCorrelationData"
}}

