% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_exposure_data_handler.r
\docType{methods}
\name{getFactorExposureDataObject}
\alias{getFactorExposureDataObject}
\alias{getFactorExposureDataObject,VirtualFactorExposureDataHandler-method}
\title{Get factor_exposure_data stored in object}
\usage{
getFactorExposureDataObject(object)


  \S4method{getFactorExposureDataObject}{VirtualFactorExposureDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualFactorExposureDataHandler"}
}
\value{
\code{factor_exposure_data} object of class "VirtualFactorExposureData"

\code{factor_exposure_data} object of class "VirtualFactorExposureData"
}
\description{
Returns factor_exposure_data object of class "VirtualFactorExposureData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualFactorExposureDataHandler}: Get factor_exposure_data stored in object

Returns factor_exposure_data object of class "VirtualFactorExposureData"
}}

