% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_variance_data_handler.r
\docType{methods}
\name{getFactorVarianceDataObject}
\alias{getFactorVarianceDataObject}
\alias{getFactorVarianceDataObject,VirtualFactorVarianceDataHandler-method}
\title{Get factor_variance stored in object}
\usage{
getFactorVarianceDataObject(object)


  \S4method{getFactorVarianceDataObject}{VirtualFactorVarianceDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualFactorVarianceDataHandler"}
}
\value{
\code{factor_variance} object of class "VirtualFactorVarianceData"

\code{factor_variance} object of class "VirtualFactorVarianceData"
}
\description{
Returns factor_variance object of class "VirtualFactorVarianceData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualFactorVarianceDataHandler}: Get factor_variance stored in object

Returns factor_variance object of class "VirtualFactorVarianceData"
}}

