% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{getForceUniqueRows}
\alias{getForceUniqueRows}
\title{Are stored rows forced to be unique}
\usage{
getForceUniqueRows(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{unique_rows} logical, are stored rows unique
}
\description{
Returns logical value indicating if stored reference data
rows are forced to be unique.
}

