% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_factor_returns_data_handler.r
\docType{methods}
\name{getImpliedFactorReturnsDataObject}
\alias{getImpliedFactorReturnsDataObject}
\alias{getImpliedFactorReturnsDataObject,VirtualImpliedFactorReturnsDataHandler-method}
\title{Get implied_factor_returns stored in object}
\usage{
getImpliedFactorReturnsDataObject(object)


  \S4method{getImpliedFactorReturnsDataObject}{VirtualImpliedFactorReturnsDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualImpliedFactorReturnsDataHandler"}
}
\value{
\code{implied_factor_returns} object of class "VirtualImpliedFactorReturnsData"

\code{implied_factor_returns} object of class "VirtualImpliedFactorReturnsData"
}
\description{
Returns implied_factor_returns object of class "VirtualImpliedFactorReturnsData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualImpliedFactorReturnsDataHandler}: Get implied_factor_returns stored in object

Returns implied_factor_returns object of class "VirtualImpliedFactorReturnsData"
}}

