% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getInputData}
\alias{getInputData}
\alias{getInputData,VirtualTransformationComputation-method}
\title{Get value of computation input}
\usage{
getInputData(object)

\S4method{getInputData}{VirtualTransformationComputation}(object)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{data} data.frame data stored in input

\code{data} data.frame data stored in input
}
\description{
Returns data from internal input datastore.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformationComputation}: Get value of computation input

Returns data from internal input datastore.
}}

