% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instrument_betas_data_handler.r
\docType{methods}
\name{getInstrumentBetasDataObject}
\alias{getInstrumentBetasDataObject}
\alias{getInstrumentBetasDataObject,VirtualInstrumentBetasDataHandler-method}
\title{Get instrument_betas stored in object}
\usage{
getInstrumentBetasDataObject(object)


  \S4method{getInstrumentBetasDataObject}{VirtualInstrumentBetasDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualInstrumentBetasDataHandler"}
}
\value{
\code{instrument_betas} object of class "VirtualInstrumentBetasData"

\code{instrument_betas} object of class "VirtualInstrumentBetasData"
}
\description{
Returns instrument_betas object of class "VirtualInstrumentBetasData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualInstrumentBetasDataHandler}: Get instrument_betas stored in object

Returns instrument_betas object of class "VirtualInstrumentBetasData"
}}

