% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_data_handler.r
\name{getMarketDataObject}
\alias{getMarketDataObject}
\title{Get market_data stored in object}
\usage{
getMarketDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualMarketDataHandler"}
}
\value{
\code{market_data} object of class "VirtualMarketData"
}
\description{
Returns market_data object of class "VirtualMarketData"
}

