% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_style_data_handler.r
\docType{methods}
\name{getMarketStyleDataObject}
\alias{getMarketStyleDataObject}
\alias{getMarketStyleDataObject,VirtualMarketStyleDataHandler-method}
\title{Get market_style stored in object}
\usage{
getMarketStyleDataObject(object)

\S4method{getMarketStyleDataObject}{VirtualMarketStyleDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualMarketStyleDataHandler"}
}
\value{
\code{market_style} object of class "VirtualMarketStyleData"

\code{market_style} object of class "VirtualMarketStyleData"
}
\description{
Returns market_style object of class "VirtualMarketStyleData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualMarketStyleDataHandler}: Get market_style stored in object

Returns market_style object of class "VirtualMarketStyleData"
}}

