% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{getNonNAColumnNames}
\alias{getNonNAColumnNames}
\alias{getNonNAColumnNames,VirtualDataSourceClient-method}
\title{Get non - NA's column names}
\usage{
getNonNAColumnNames(object)

\S4method{getNonNAColumnNames}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{non_na_cols} character vector with list of returned column names

\code{non_na_cols} character vector with list of returned column names
}
\description{
Returns column names that are required to have non-na values
rows where these columns have NA's will be removed
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Get non - NA's column names

Returns column names that are required to have non-na values
rows where these columns have NA's will be removed
}}

