% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{getOutputData}
\alias{getOutputData}
\title{Get value of computation output}
\usage{
getOutputData(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}
}
\value{
\code{data} data.frame data stored in output
}
\description{
Returns data from internal output datastore.
}

