% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{getOutputVariablesNames}
\alias{getOutputVariablesNames}
\alias{getOutputVariablesNames,VirtualBaseTransformation-method}
\title{Get names of output columns}
\usage{
getOutputVariablesNames(object)

\S4method{getOutputVariablesNames}{VirtualBaseTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{output_colnms} character vector of names of output columns

\code{output_colnms} character vector of names of output columns
}
\description{
Returns Names of the variables that are returned by computation object.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualBaseTransformation}: Get names of output columns

Returns Names of the variables that are returned by computation object.
}}

