% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_data_handler.r
\docType{methods}
\name{getPortfolioDataObject}
\alias{getPortfolioDataObject}
\alias{getPortfolioDataObject,VirtualPortfolioDataHandler-method}
\title{Get portfolio stored in object}
\usage{
getPortfolioDataObject(object)

\S4method{getPortfolioDataObject}{VirtualPortfolioDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualPortfolioDataHandler"}
}
\value{
\code{portfolio} object of class "Portfolio"

\code{portfolio} object of class "Portfolio"
}
\description{
Returns portfolio object of class "Portfolio"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualPortfolioDataHandler}: Get portfolio stored in object

Returns portfolio object of class "Portfolio"
}}

