% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_data_handler.r
\name{getPositionDataObject}
\alias{getPositionDataObject}
\title{Get position_data stored in object}
\usage{
getPositionDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualPositionDataHandler"}
}
\value{
\code{position_data} object of class "VirtualPositionData"
}
\description{
Returns position_data object of class "VirtualPositionData"
}

