% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppmodel_client.r
\docType{methods}
\name{getPPModelClass}
\alias{getPPModelClass}
\alias{getPPModelClass,VirtualPPModelClient-method}
\title{Get Post Processing Model Class}
\usage{
getPPModelClass(object)

\S4method{getPPModelClass}{VirtualPPModelClient}(object)
}
\arguments{
\item{object}{object of class "VirtualPPModelClient"}
}
\value{
\code{model_class} "character" name of the model class

\code{model_class} "character" name of the model class
}
\description{
Returns post processing model used to assembly data.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualPPModelClient}: Get Post Processing Model Class

Returns post processing model used to assembly data.
}}

