% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_data_handler.r
\docType{methods}
\name{getPriceDataObject}
\alias{getPriceDataObject}
\alias{getPriceDataObject,VirtualPriceDataHandler-method}
\title{Get price_data stored in object}
\usage{
getPriceDataObject(object)

\S4method{getPriceDataObject}{VirtualPriceDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualPriceDataHandler"}
}
\value{
\code{price_data} object of class "VirtualPriceData"

\code{price_data} object of class "VirtualPriceData"
}
\description{
Returns price_data object of class "VirtualPriceData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualPriceDataHandler}: Get price_data stored in object

Returns price_data object of class "VirtualPriceData"
}}

