% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{getReferenceData}
\alias{getReferenceData}
\alias{getReferenceData,VirtualReferenceData-method}
\title{get stored data}
\usage{
getReferenceData(object)

\S4method{getReferenceData}{VirtualReferenceData}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{data} data.frame with stored data..

\code{data} data.frame with stored data..
}
\description{
Returns data.frame with stored data.
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceData}: get stored data

Returns data.frame with stored data.
}}

