% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.R
\name{getReferenceData}
\alias{getReferenceData}
\title{get stored data}
\usage{
getReferenceData(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{data} data.frame with stored data..
}
\description{
Returns data.frame with stored data.
}

