% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\docType{methods}
\name{getReferenceDataObject}
\alias{getReferenceDataObject}
\alias{getReferenceDataObject,VirtualReferenceDataTransformation-method}
\title{Get stored VirtualReferenceData Object}
\usage{
getReferenceDataObject(object)

\S4method{getReferenceDataObject}{VirtualReferenceDataTransformation}(object)
}
\arguments{
\item{object}{object of class 'VirtualReferenceDataTransformation'.}
}
\value{
\code{ref_data} object of class 'VirtualReferenceData'.

\code{ref_data} object of class 'VirtualReferenceData'.
}
\description{
Returns VirtualReferenceData object on which transformation is beeing done
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualReferenceDataTransformation}: Get stored VirtualReferenceData Object

Returns VirtualReferenceData object on which transformation is beeing done
}}

