% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referenceobject.r
\name{getRequiredVariablesNames}
\alias{getRequiredVariablesNames}
\title{Get names of required variables stored in data}
\usage{
getRequiredVariablesNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceObject'.}
}
\value{
\code{required_colnms} character vector, list of required column names
}
\description{
Get names of required variables stored in data
}

