% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_handler.r
\docType{methods}
\name{getRiskModelObject}
\alias{getRiskModelObject}
\alias{getRiskModelObject,VirtualRiskModelHandler-method}
\title{Get Risk Model Object}
\usage{
getRiskModelObject(object)

\S4method{getRiskModelObject}{VirtualRiskModelHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelHandler'.}
}
\value{
\code{risk_model} object of class "VirtualRiskModel".

\code{risk_model} object of class "VirtualRiskModel".
}
\description{
Returns risk model object
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualRiskModelHandler}: Get Risk Model Object

Returns risk model object
}}

