% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model_objectstore_client.r
\name{getRiskModelObjectstoreComponentName}
\alias{getRiskModelObjectstoreComponentName}
\title{Get Risk Model Component Name}
\usage{
getRiskModelObjectstoreComponentName(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModelObjectstoreClient'.}
}
\value{
\code{component} 'character', name of the component beeing accessed.
}
\description{
Returns name of the component that given class is accessing
}

