% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risk_model.r
\name{getRiskModelUniverse}
\alias{getRiskModelUniverse}
\title{Get Risk Model Universe}
\usage{
getRiskModelUniverse(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualRiskModel'.}
}
\value{
\code{model_universe} 'character', name of the universe covered by the model
}
\description{
Returns name of universe for which model is computed.
}

