% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\name{getStartDate}
\alias{getStartDate}
\title{Get start date of portfolio timespan}
\usage{
getStartDate(object, ...)
}
\arguments{
\item{object}{object of class 'Portfolio'.}
}
\value{
\code{start} Date start of data range.
}
\description{
Get start date of portfolio timespan
}

