% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{getStoredNRows}
\alias{getStoredNRows}
\title{Number of currently stored rows}
\usage{
getStoredNRows(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{stored_rows} integer, list of stored column names
}
\description{
Returns count of currently stored rows
}

