% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{getStoredVariablesNames}
\alias{getStoredVariablesNames}
\title{Get names of stored reference data variables}
\usage{
getStoredVariablesNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}
}
\value{
\code{column_names} character vector, list of stored column names
}
\description{
Returns vector of column names of stored data
}

