% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_data_handler.r
\name{getStrategyDataObject}
\alias{getStrategyDataObject}
\title{Get strategy_data stored in object}
\usage{
getStrategyDataObject(object, ...)
}
\arguments{
\item{object}{object of class "VirtualStrategyDataHandler"}
}
\value{
\code{strategy_data} object of class "VirtualStrategyData"
}
\description{
Returns strategy_data object of class "VirtualStrategyData"
}

