% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trade_data_handler.r
\docType{methods}
\name{getTradeDataObject}
\alias{getTradeDataObject}
\alias{getTradeDataObject,VirtualTradeDataHandler-method}
\title{Get trade_data stored in object}
\usage{
getTradeDataObject(object)

\S4method{getTradeDataObject}{VirtualTradeDataHandler}(object)
}
\arguments{
\item{object}{object of class "VirtualTradeDataHandler"}
}
\value{
\code{trade_data} object of class "VirtualTradeData"

\code{trade_data} object of class "VirtualTradeData"
}
\description{
Returns trade_data object of class "VirtualTradeData"
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTradeDataHandler}: Get trade_data stored in object

Returns trade_data object of class "VirtualTradeData"
}}

