% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategy_portfolio.r
\docType{methods}
\name{getTraderID}
\alias{getTraderID}
\alias{getTraderID,StrategyPortfolio-method}
\title{Returns TraderID}
\usage{
getTraderID(object)

\S4method{getTraderID}{StrategyPortfolio}(object)
}
\arguments{
\item{object}{object of class 'StrategyPortfolio'.}
}
\value{
\code{trader_id} integer.
}
\description{
Returns TraderID
}
\section{Methods (by class)}{
\itemize{
\item \code{StrategyPortfolio}: Returns TraderID
}}

