% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_values_data_handler.r
\docType{methods}
\name{getValueName}
\alias{getValueName}
\alias{getValueName,VirtualAvailableValuesDataHandler-method}
\title{Getter for the value column name}
\usage{
getValueName(object)

\S4method{getValueName}{VirtualAvailableValuesDataHandler}(object)
}
\arguments{
\item{object}{object of class 'VirtualAvailableValuesDataHandler'.}
}
\value{
\code{name} "character"
}
\description{
Getter for the value column name
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualAvailableValuesDataHandler}: Getter for the value column name
}}

