% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\name{hasFactorColumnNames}
\alias{hasFactorColumnNames}
\title{Are there any columns to  factorize ?}
\usage{
hasFactorColumnNames(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_factorized_cols} logical
}
\description{
Returns TRUE if there are some columns to factorize
}

