% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasource_client.r
\docType{methods}
\name{hasNonNAColumnNames}
\alias{hasNonNAColumnNames}
\alias{hasNonNAColumnNames,VirtualDataSourceClient-method}
\title{Are any non - NA's column names defined in object}
\usage{
hasNonNAColumnNames(object)

\S4method{hasNonNAColumnNames}{VirtualDataSourceClient}(object)
}
\arguments{
\item{object}{object of class 'VirtualDataSourceClient'.}
}
\value{
\code{has_non_na_cols} logical

\code{has_non_na_cols} logical
}
\description{
Returns true if object has columns that are required to have non-na-values
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualDataSourceClient}: Are any non - NA's column names defined in object

Returns true if object has columns that are required to have non-na-values
}}

