% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{isComputed}
\alias{isComputed}
\title{Is computation done}
\usage{
isComputed(object, ...)
}
\arguments{
\item{object}{object of class 'VirtualBaseTransformation'.}
}
\value{
\code{is_computed} logical inicating if computation is done
}
\description{
Returns logical value indicating if computation has
been done.
}

