% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations_handler.r
\name{isTransformationAttached}
\alias{isTransformationAttached}
\title{Check if given transformation is attached}
\usage{
isTransformationAttached(object, transformation, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationHandler'.}

\item{transformation}{object of class "VirtualReferenceDataTransformation"}
}
\value{
\code{is_attached} logical.
}
\description{
Returns logical value indicating if given transformation
is attached to the object
}

