% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{setComputation}
\alias{setComputation}
\alias{setComputation,VirtualTransformation,VirtualTransformationComputation-method}
\title{Set computation object}
\usage{
setComputation(object, computation)


  \S4method{setComputation}{VirtualTransformation,VirtualTransformationComputation}(object,
  computation)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{computation}{object of type VirtualTransformationComputation}
}
\value{
\code{object} object of class 'VirtualTransformation'

\code{object} object of class 'VirtualTransformation'
}
\description{
sets computation object encapsulated in Transformation.
Handles all checks and settings.
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformation,computation = VirtualTransformationComputation}: Set computation object

sets computation object encapsulated in Transformation.
Handles all checks and settings.
}}

