% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{setComputationInput}
\alias{setComputationInput}
\alias{setComputationInput,VirtualTransformation,VirtualReferenceData-method}
\title{Set value of computation input}
\usage{
setComputationInput(object, data)


  \S4method{setComputationInput}{VirtualTransformation,VirtualReferenceData}(object,
  data)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{data}{object of class 'VirtualReferenceData'.}
}
\value{
\code{object} object of class 'VirtualTransformation'

\code{object} object of class 'VirtualTransformation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualTransformation,data = VirtualReferenceData}: Set value of computation input

Sets internal input datastore to new data and performs data validity checks .
}}

