% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\docType{methods}
\name{setEndDate}
\alias{setEndDate}
\alias{setEndDate,Portfolio-method}
\title{Set end date of portfolio timespan}
\usage{
setEndDate(object, end)

\S4method{setEndDate}{Portfolio}(object, end)
}
\arguments{
\item{object}{object of class 'Portfolio'.}

\item{end}{Date end of data range.}
}
\value{
\code{object} object of class 'Portfolio'.

\code{object} object of class 'Portfolio'.
}
\description{
Set end date of portfolio timespan
}
\section{Methods (by class)}{
\itemize{
\item \code{Portfolio}: Set end date of portfolio timespan
}}

