% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\name{setInputData}
\alias{setInputData}
\title{Set value of computation input}
\usage{
setInputData(object, data, ...)
}
\arguments{
\item{object}{object of class 'VirtualTransformationComputation'.}

\item{data}{data to store in.}
}
\value{
\code{object} object of class 'VirtualTransformationComputation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}

