% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\docType{methods}
\name{setReferenceData}
\alias{setReferenceData}
\alias{setReferenceData,VirtualReferenceData,data.frame-method}
\title{Set stored data to value provided}
\usage{
setReferenceData(object, data)

\S4method{setReferenceData}{VirtualReferenceData,data.frame}(object, data)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data has to comply with required variable names.}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.

\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Sets new values to stored data. Erases any previous data without warning
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualReferenceData,data = data.frame}: Set stored data to value provided

Sets new values to stored data. Erases any previous data without warning
}}

