% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r, R/referencedata_transformation.r
\name{setReferenceData}
\alias{setReferenceData}
\title{Set stored data to value provided}
\usage{
setReferenceData(object, ref_data, ...)

setReferenceData(object, ref_data, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{ref_data}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data has to comply with required variable names.}

\item{object}{object of class 'VirtualReferenceDataTransformation'.}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.

\code{ref_data} object of class 'VirtualReferenceDataTransformation'.
}
\description{
Sets new values to stored data. Erases any previous data without warning

Sets VirtualReferenceData object on which transformation is beeing done
}

