% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata_transformation.r
\name{setReferenceDataObject}
\alias{setReferenceDataObject}
\title{Set VirtualReferenceData Object}
\usage{
setReferenceDataObject(object, ref_data, ...)
}
\arguments{
\item{object}{object of class 'VirtualReferenceDataTransformation'.}

\item{ref_data}{object of class 'VirtualReferenceData'.}
}
\value{
\code{ref_data} object of class 'VirtualReferenceDataTransformation'.
}
\description{
Sets VirtualReferenceData object on which transformation is beeing done
}

