% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio.r
\name{setStartDate}
\alias{setStartDate}
\title{Set start date of portfolio timespan}
\usage{
setStartDate(object, start, ...)
}
\arguments{
\item{object}{object of class 'Portfolio'.}

\item{start}{Date start of data range.}
}
\value{
\code{object} object of class 'Portfolio'.
}
\description{
Set start date of portfolio timespan
}

