% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/virtual_transformation.r
\docType{methods}
\name{triggerComputation}
\alias{triggerComputation}
\alias{triggerComputation,VirtualTransformation-method}
\title{Trigger transformation Computation}
\usage{
triggerComputation(object, force = FALSE)

\S4method{triggerComputation}{VirtualTransformation}(object, force = FALSE)
}
\arguments{
\item{object}{object of class 'VirtualTransformation'.}

\item{force}{logical value indicating if new computation has to be forced if it has been
computed before.}
}
\value{
\code{object} object of class 'VirtualTransformation'

\code{object} object of class 'VirtualTransformation'
}
\description{
Sets internal input datastore to new data and performs data validity checks .
}
\section{Methods (by class)}{
\itemize{
\item \code{VirtualTransformation}: Trigger transformation Computation

Sets internal input datastore to new data and performs data validity checks .
}}

