% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/referencedata.r
\name{updateVariables}
\alias{updateVariables}
\title{Update values in specified columns}
\usage{
updateVariables(object, data, var.names)
}
\arguments{
\item{object}{object of class 'VirtualReferenceData'.}

\item{data}{data.frame with new data has to comply with required variable names.}

\item{var.names}{column names of data in both incomming data and stored data}
}
\value{
\code{object} object object of class 'VirtualReferenceData'.
}
\description{
Sets new values to specified columns in stored data.
The number of rows of incoming data has to be the same as number of
rows currently stored.
}

