% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validator.r
\docType{methods}
\name{validateData}
\alias{validateData}
\alias{validateData,VirtualDataValidator,data.frame-method}
\title{Apply validation to dataframe}
\usage{
validateData(object, data)

\S4method{validateData}{VirtualDataValidator,data.frame}(object, data)
}
\arguments{
\item{object}{object of class 'VirtualDataValidator'.}
}
\value{
\code{data} data.frame of validated data
}
\description{
Public method to return dataframe post validation
or halt execution of halt is TRUE
}
\section{Methods (by class)}{
\itemize{
\item \code{object = VirtualDataValidator,data = data.frame}: Apply validation to dataframe

Public method to return dataframe post validation
or halt execution of halt is TRUE
}}

