% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodbc_client.r
\docType{class}
\name{VirtualSQLQuery-class}
\alias{VirtualSQLQuery-class}
\alias{prepareSQLQuery}
\title{Virtual S4 class sql queries handling.}
\usage{
prepareSQLQuery(object, key_values, ...)
}
\arguments{
\item{object}{object of class 'VirtualSQLQuery'.}
}
\value{
\code{object} object of class 'VirtualSQLQuery'.
}
\description{
Implements handling of access to data via SQL queries.

Parses key_values to vector of SQL query strings
}
\section{Slots}{

\describe{
\item{\code{db_schema}}{= "character",}

\item{\code{sql_query}}{= "character",}

\item{\code{key_cols}}{= "character",}

\item{\code{key_values}}{= "data.frame",}

\item{\code{query_parser}}{= "function",}

\item{\code{results_parser}}{= "function"}
}}

